Refactor settings translations into separate files per language
- Split `src/data/settings-translations.ts` into `src/data/settings-translations/id.ts`, `src/data/settings-translations/en.ts`, and `src/data/settings-translations/index.ts`.
- Added a test file `src/data/settings-translations/settings-translations.test.ts` to verify key consistency and correct exports.
- Maintained backward compatibility for imports.
