--- src/data/settings-translations.ts
@@ -1,1023 +1,0 @@
-/**
- * Settings Page Translations
- * Localized text for settings page in Indonesian and English
- */
-
-import { APP_CONFIG } from "@/config/app-config";
-
-export const SETTINGS_TRANSLATIONS = {
-  id: {
-    // Header
-    title: "Pengaturan",
-
-    // Stats Section
-    statsLabel: "Statistik Ibadah",
-    statsComingSoon: "Segera Hadir",
-    statsDescription: "Matangkan niat, pantau hasil nanti ✨",
-
-    // Location Section
-    locationLabel: "Lokasi",
-    locationUpdating: "Memperbarui...",
-    locationDetecting: "Mendeteksi...",
-
-    // Calculation Method
-    methodLabel: "Metode",
-    methodPlaceholder: "Pilih metode",
-
-    // Prayer Notifications
-    notificationTitle: "Notifikasi Adzan",
-    notificationButton: "Aktifkan Notifikasi Adzan",
-    notificationPermissionTitle: "Aktifkan Notifikasi Sholat",
-    notificationPermissionDesc: "Untuk mendapatkan pengingat waktu sholat yang akurat, mohon izinkan akses notifikasi pada perangkat Anda.",
-    notificationPermissionButton: "Izinkan Akses Notifikasi",
-    notificationDenied: "Izin notifikasi ditolak. Mohon aktifkan di pengaturan browser Anda.",
-    prayerNotifications: "Notifikasi Waktu Sholat",
-    notificationsEnabled: "Notifikasi aktif",
-    notificationsDisabled: "Notifikasi nonaktif",
-    selectPrayerTimes: "Pilih Waktu Sholat",
-    notificationLocationNote: "Notifikasi akan dikirim berdasarkan lokasi yang tersimpan di aplikasi.",
-
-    // Prayer Names
-    fajr: "Subuh",
-    dhuhr: "Dzuhur",
-    asr: "Ashar",
-    maghrib: "Maghrib",
-    isha: "Isya",
-
-    // Theme
-    themeTitle: "Tampilan Aplikasi",
-
-    // Audio Settings
-    audioTitle: "Pengaturan Audio",
-    muadzinLabel: "Suara Adzan",
-
-    // Language Settings
-    languageTitle: "Bahasa",
-    languageDescription: "Bahasa aplikasi akan berubah sesuai pilihan.",
-
-    // Support Card
-    supportTitle: "Dukung Nawaetu",
-    supportPremiumText: "Terima kasih telah menjadi Muhsinin! Hasil infaq digunakan untuk server & pengembangan.",
-    supportText: "Bantu kami menjaga aplikasi tetap gratis dan bebas iklan selamanya.",
-    infaqButtonPremium: "Infaq Lagi",
-    infaqButton: "Infaq",
-
-    // About Section
-    aboutAppName: "Nawaetu",
-    aboutTagline: "Luruskan Niat, Sempurnakan Ibadah",
-    aboutDescription: "Teman ibadah digital yang menjaga setiap langkahmu tetap sesuai tuntunan Al-Qur'an & Sunnah.",
-    aboutHashtag: "#LuruskanNiat",
-    aboutVersion: `v${APP_CONFIG.version} • ${APP_CONFIG.status}`,
-    aboutTapForInfo: "Ketuk untuk info lengkap",
-    aboutWhatsNew: "Apa yang Baru",
-    aboutUpdate1: "Riwayat Chat AI tersimpan antar sesi",
-    aboutUpdate2: "Rate Limit cerdas (Free 5x/hari, Muhsinin 25x/hari)",
-    aboutUpdate3: "Update Respon AI (Lebih ringkas & tanpa basa-basi)",
-    aboutUpdate4: "Perbaikan UI Navigasi & Dark Mode",
-    aboutUpdate5: "Optimasi performa & perbaikan bug",
-    aboutFeature1: "Al-Qur'an",
-    aboutFeature2: "Waktu Sholat",
-    aboutFeature3: "Mentor AI",
-    aboutFeature4: "Multi-Bahasa",
-    aboutWhatIsTitle: "Apa itu Nawaetu?",
-    aboutApproachTitle: "Pendekatan Kami",
-    aboutApproachDesc: "Kami menggabungkan panduan ibadah dengan gamifikasi (Misi & Streak) serta asisten AI untuk membangun kebiasaan yang istiqomah.",
-    aboutLastUpdate: `Diperbarui: ${APP_CONFIG.lastUpdated}`,
-
-    // Profile Dialog
-    profileGenderLabel: "Jenis Kelamin",
-    profileMale: "Laki-laki",
-    profileFemale: "Perempuan",
-    profileGenderHint: "Pilih untuk menyesuaikan misi dan fitur",
-    profileArchetypeLabel: "Fokus Ibadah (Tipe Pejuang)",
-    profileArchetypeBeginner: "Pemula",
-    profileArchetypeMobilizer: "Penggerak",
-    profileArchetypeWarrior: "Mujahid",
-    profileProgressLabel: "Progress Level",
-    profileSupport: "Dukung",
-    profileAddDonation: "Tambah Infaq",
-    profileTitleUnlock: "Gelar Terbuka",
-    profileTitleLocked: "Terkunci (Butuh Lv.{level})",
-    profilePhilosophy: "Filosofi Gelar",
-    profileRewards: "Privilege & Rewards",
-    profileUseTitle: "Pasang Gelar Ini",
-    profileCurrentTitle: "Sedang Digunakan",
-    profileMinimalLevel: "Minimal Level {level}",
-    profileSave: "Simpan Profil",
-    profileTitleLabel: "Gelar Ibadah",
-    profileNameLabel: "Nama Lengkap",
-    profileAppSettings: "Pengaturan Aplikasi",
-    profileSyncData: "Sinkronisasi Data",
-    profileSyncing: "Menyinkronkan...",
-    profileShareApp: "Bagikan Aplikasi",
-    profileLogout: "Keluar",
-    profileLogoutConfirm: "Yakin ingin keluar akun?",
-    profileLogoutConfirmYes: "Ya, Keluar",
-    profileLogoutConfirmNo: "Batal",
-    profileGuestMode: "Mode Guest",
-    profileFooter: "Terima kasih telah menggunakan Nawaetu",
-    profileAuthTitle: "Simpan Progress Ibadahmu",
-    profileAuthDesc: "Login untuk menyimpan data streak, bookmark, dan status donatur secara permanen di cloud.",
-    profileAuthButton: "Login dengan Google",
-    profileRoleGuest: "Pengguna Setia",
-    profileRolePremium: "Muhsinin Nawaetu",
-    profileStreakLabel: "Hari Streak",
-    profileTotalDays: "Total Hari",
-
-    // Streak & Progress
-    profileNeedXP: "Butuh {xp} XP lagi untuk naik ke Level {level}",
-    profileStreak: "Streak",
-    profileRecord: "Rekor",
-    profileDays: "Hari",
-    profileStreakMaintain: "Pertahankan istiqamahmu setiap hari!",
-    profileBestAchievement: "Pencapaian istiqamah terbaikmu.",
-    profileTitleCollection: "Koleksi Gelar",
-
-    // XP Earning Methods Section
-    xpSectionTitle: "Cara Mendapatkan XP",
-    xpMethodReadQuran: "Baca Quran",
-    xpReadQuranReward: "+5 XP/ayat",
-    xpMethodTasbih: "Tasbih",
-    xpTasbihReward: "+50 XP/sesi",
-    xpMethodDailyMission: "Misi Harian",
-    xpDailyMissionReward: "+25-100 XP/misi",
-    xpMethodStatistics: "Statistik",
-    xpStatisticsHint: "Lihat Progress",
-
-    // Navigation Menu
-    navHome: "Beranda",
-    navQuran: "Al-Qur'an",
-    navTasbih: "Tasbih",
-    navQibla: "Kiblat",
-    navSettings: "Atur",
-
-    // Qibla Page
-    qiblaFinding: "TENANG, CARI KIBLAT",
-    qiblaAligned: "HATI TENANG",
-    qiblaAlignedBadge: "KIBLAT TEPAT",
-    qiblaPermissionTitle: "Izin Akses Kompas",
-    qiblaPermissionDesc: "Nawaetu memerlukan akses sensor gerak HP Anda untuk menentukan arah kiblat dengan presisi.",
-    qiblaPermissionButton: "Aktifkan Sensor",
-    qiblaLocationError: "Izin lokasi diperlukan untuk menghitung arah kiblat.",
-    qiblaGeoError: "Geolocation tidak didukung di browser ini.",
-    qiblaCompassDenied: "Izin kompas ditolak.",
-    qiblaCompassFailed: "Gagal meminta izin kompas.",
-    qiblaDistance: "Jarak ke Ka'bah: {distance} km",
-    qiblaSearching: "Mencari lokasi...",
-    qiblaDalilTitle: "Dalil Arah Kiblat",
-    qiblaDalilQuranRef: "QS. Al-Baqarah 2:144",
-    qiblaDalilQuranText: "Hadapkan wajah ke Masjidil Haram.",
-    qiblaDalilContextTitle: "Asbabun nuzul:",
-    qiblaDalilContextText: "Peralihan kiblat dari Baitul Maqdis ke Ka'bah.",
-
-    // Tasbih Page
-    tasbihTitle: "Tasbih Digital",
-    tasbihSubtitle: "Zikir Penenang Hati",
-    tasbihSelectZikir: "Pilih Zikir",
-    tasbihTarget: "Target",
-    tasbihDaily: "Harian",
-    tasbihStreak: "Streak",
-    tasbihReset: "Reset",
-    tasbihResetConfirm: "Reset hitungan?",
-    tasbihTap: "Tap",
-    tasbihDays: "Hari",
-    tasbihListTitle: "Daftar Bacaan",
-    tasbihVibrate: "Getar",
-    tasbihSound: "Suara",
-    tasbihDual: "Dual",
-    tasbihMute: "Mute",
-    tasbihFreeMode: "Mode Bebas",
-    tasbihCounterLabel: "Counter",
-    tasbihComplete: "Alhamdulillah!",
-    tasbihCompleteMessage: "Target zikir hari ini tercapai. Semoga berkah!",
-    tasbihContinue: "Lanjut Zikir",
-    tasbihBack: "Kembali",
-    tasbihNextUp: "Lanjut ke",
-    tasbihRepeat: "Ulangi",
-    tasbihRepeatReading: "Ulangi Bacaan",
-    tasbihPresetTasbihLabel: "Tasbih",
-    tasbihPresetTasbihLatin: "Subhanallah",
-    tasbihPresetTasbihTadabbur: "Maha Suci Allah dari segala kekurangan.",
-    tasbihPresetTahmidLabel: "Tahmid",
-    tasbihPresetTahmidLatin: "Alhamdulillah",
-    tasbihPresetTahmidTadabbur: "Segala puji bagi Allah atas segala nikmat.",
-    tasbihPresetTakbirLabel: "Takbir",
-    tasbihPresetTakbirLatin: "Allahu Akbar",
-    tasbihPresetTakbirTadabbur: "Allah Maha Besar.",
-    tasbihPresetIstighfarLabel: "Istighfar",
-    tasbihPresetIstighfarLatin: "Astaghfirullah",
-    tasbihPresetIstighfarTadabbur: "Aku memohon ampunan kepada Allah SWT.",
-    tasbihPresetSholawatJibrilLabel: "Sholawat Jibril",
-    tasbihPresetSholawatJibrilLatin: "Shallallahu 'ala Muhammad",
-    tasbihPresetSholawatJibrilTadabbur: "Semoga Allah melimpahkan rahmat kepada Nabi Muhammad.",
-
-    // Quran Page
-    quranTitle: "Al-Qur'an Digital",
-    quranSubtitle: "Lebih dari sekedar membaca",
-    quranChooseSurah: "Pilih Surah",
-    quranBookmark: "Bookmark",
-    quranShare: "Bagikan",
-    quranCopy: "Salin",
-    quranTafsir: "Tafsir",
-    quranAudio: "Audio",
-
-    // Home Page
-    homeGreetingMorning: "Selamat Pagi",
-    homeGreetingNoon: "Selamat Siang",
-    homeGreetingAfternoon: "Selamat Sore",
-    homeGreetingEvening: "Selamat Malam",
-    homeGreetingWelcome: "Selamat Datang",
-    homeLocation: "Lokasi",
-    homeLoading: "Loading...",
-
-    // Home Widgets
-    widgetLastRead: "Terakhir Dibaca",
-    widgetContinueReading: "Lanjut Baca",
-    widgetMissions: "Misi Hari Ini",
-    widgetMissionsToday: "Misi Hari Ini",
-    widgetNoMissions: "Belum ada misi hari ini",
-    widgetPrayerTimes: "Jadwal Sholat",
-    widgetNextPrayer: "Sholat Berikutnya",
-    widgetQuoteOfDay: "Quote Hari Ini",
-    widgetRamadhan: "Ramadhan",
-    homeAiTitle: "Tanya Nawaetu",
-    homeAiSubtitle: "Asisten Muslim AI",
-    homeNextLabel: "Selanjutnya",
-    homePreparation: "Persiapan",
-    homeTowards: "Menuju {prayer}",
-    homeAdzanNow: "Adzan {prayer}",
-    homeAdzanMessage: "Selamat menunaikan sholat",
-    homePrayerTimesToday: "Jadwal Sholat Hari Ini",
-    homeFindMosque: "Cari Masjid",
-    homeFindMosqueAria: "Cari Masjid Terdekat via Google Maps",
-
-    // Mosque Finder Modal
-    mosqueFinderTitle: "Mau Sholat di Mana?",
-    mosqueFinderDesc: "Pilih jenis tempat sholat yang kamu butuhkan:",
-    mosqueFinderOptionNearby: "Masjid Terdekat",
-    mosqueFinderOptionMusholla: "Musholla",
-    mosqueFinderOptionGrand: "Masjid Raya",
-    mosqueFinderOptionRestArea: "Rest Area",
-    mosqueFinderButton: "Buka Google Maps",
-    homeLocationRequiredTitle: "Izin Lokasi Diperlukan",
-    homeSetLocationNow: "Set Lokasi",
-    homeLocationDefaultTitle: "Lokasi Default (Jakarta)",
-    homeLocationDefaultDesc: "Jadwal sholat saat ini berdasarkan Jakarta. Ketuk ikon lokasi di atas untuk memperbarui ke lokasi Anda.",
-    homeLocationRequiredDesc: "Nawaetu membutuhkan lokasi Anda untuk menghitung jadwal sholat dan arah kiblat yang akurat.",
-    homeEnableLocation: "Aktifkan Lokasi",
-    homeLocationHint: "Browser akan meminta izin setelah tombol diklik.",
-    homeLastReadStartTitle: "Mulai Tilawah",
-    homeLastReadStartSubtitle: "Baca Al-Quran hari ini",
-    homeLastReadLabel: "Terakhir Baca",
-    homeMissionFocusTitle: "Fokus Harian",
-    homeMissionDailyTarget: "Fokus utama hari ini",
-    homeMissionCompleted: "Selesai",
-    homeMissionViewAll: "Lihat Semua Misi",
-    homeMissionSelectGenderHint: "💡 Pilih jenis kelamin di Profil untuk misi yang lebih personal",
-    homeMissionSpecial: "Special",
-    homeMissionLocked: "Terkunci",
-    homeMissionLate: "Terlewat",
-    homeMissionEarly: "Awal Waktu",
-    homeMissionEarlyTitle: "Keutamaan Awal Waktu",
-    homeMissionEarlyQuote: "\"Amalan terbaik: Shalat di awal waktu.\" (HR. Tirmidzi)",
-    homeMissionLateTitle: "Waktu Hampir Habis! ({minutes} menit)",
-    homeMissionLateQuote: "\"Janganlah kalian lalai terhadap shalat.\" (QS. Al-Ma'un: 5)",
-    homeMissionListTitle: "🎯 Daftar Misi Lengkap",
-    homeMissionLatePrayerTitle: "Sholat Terlewat (Qadha)",
-    homeMissionEarlyPrayerTitle: "Keutamaan Sholat Awal Waktu",
-    homeMissionLockedFallback: "Terkunci",
-    homeMissionCompleteLate: "Catat Amalan (Terlewat)",
-    homeMissionComplete: "Selesaikan Misi",
-    homeMissionCompletedLabel: "Misi Sudah Selesai",
-    homeMissionUndoConfirm: "Ya, Batalkan",
-    homeMissionUndoCancel: "Batal",
-    homeMissionUndoPrompt: "Batalkan Penyelesaian?",
-    homeMissionLateWarningQuote: "\"Maka celakalah bagi orang-orang yang shalat, (yaitu) orang-orang yang lalai dari shalatnya.\" (QS. Al-Ma'un: 4-5)",
-    homeMissionLateWarningDesc: "Segera tunaikan sholatmu meskipun waktunya sudah lewat. Lebih baik terlambat daripada tidak sama sekali.",
-    homeMissionLateNotice: "Waktu utama misi ini sudah lewat, namun Kamu tetap bisa memvalidasi jika sudah mengamalkannya.",
-    homeMissionEarlyPraiseQuote: "\"Amalan yang paling dicintai Allah adalah sholat pada waktunya.\" (HR. Bukhari & Muslim)",
-    ramadhanHeading: "Menuju Ramadhan",
-    ramadhanDaysLeft: "Hari Lagi",
-    ramadhanReady: "Ramadhan Ready! 🌙",
-    ramadhanLevel75: "Pejuang Istiqomah",
-    ramadhanLevel50: "Siap Melangkah",
-    ramadhanLevel25: "Niat Terpasang",
-    ramadhanLevel0: "Newbie Ramadhan",
-    ramadhanPreparationLabel: "Persiapan",
-    ramadhanInfoTitle: "Persiapan Ramadhan",
-    ramadhanInfoDesc: "Persentase \"Persiapan\" dihitung dari penyelesaian misi-misi khusus di bulan Sya'ban (seperti Puasa Sunnah, Baca Quran, dll).",
-    ramadhanImproveTitle: "🎯 Cara Meningkatkan:",
-    ramadhanImproveItem1: "Selesaikan misi Puasa Sunnah Sya'ban",
-    ramadhanImproveItem2: "Lakukan Sedekah Subuh",
-    ramadhanImproveItem3: "Mulai rutinkan baca Al-Quran",
-    ramadhanUnderstand: "Mengerti",
-    streakNextMilestone: "{days} hari lagi ke {label}",
-    streakLongest: "Streak terpanjang!",
-    streakDaysLabel: "hari",
-
-    // PWA Install Prompt
-    pwaInstallTitle: "Install Nawaetu",
-    pwaInstallDesc: "Akses lebih cepat, hemat kuota, dan fitur full screen tanpa perlu download di Store.",
-    pwaInstallIosStep1: "Tap tombol Share",
-    pwaInstallIosStep2: "Pilih Add to Home Screen",
-    pwaInstallButton: "Install sekarang",
-    pwaUpdateAvailableTitle: "Update Tersedia",
-    pwaUpdateAvailableDesc: "Versi baru Nawaetu telah tersedia.",
-    pwaUpdateAction: "Update Sekarang",
-
-
-    // Mission List Modal Tabs
-    missionTabAll: "Semua",
-    missionTabDaily: "📋 Harian",
-    missionTabWeekly: "📅 Mingguan",
-    missionTabRamadhan: "🌙 Ramadhan",
-    missionTabSyaban: "🌙 Sya'ban",
-    missionTabSeasonal: "🌙 Musiman",
-    missionEmptyCategory: "Belum ada misi di kategori ini.",
-    missionEmptySeasonalTitle: "Belum ada misi khusus bulan ini.",
-    missionEmptySeasonalDesc: "Nantikan misi spesial Ramadhan segera!",
-
-    // Mission Completion Options
-    missionCompletionAlone: "Sholat Sendiri",
-    missionCompletionCongregation: "Berjamaah di Masjid",
-
-    // Mission Hukum Labels
-    hukumWajib: "Wajib",
-    hukumSunnah: "Sunnah",
-    hukumMubah: "Mubah",
-    hukumMakruh: "Makruh",
-    hukumHaram: "Haram",
-
-    // Quran Browser
-    quranLastRead: "Terakhir Dibaca",
-    quranVerse: "Ayat",
-    quranNoHistory: "Belum ada riwayat",
-    quranStartReading: "Mulai membaca sekarang.",
-    quranBookmarks: "Tanda Baca",
-    quranVersesSaved: "ayat tersimpan",
-
-    // Bookmarks Page
-    bookmarksTitle: "Koleksi Tanda Baca",
-    bookmarksSubtitle: "Catatan dan ayat pilihan Anda",
-    bookmarksEmptyTitle: "Belum ada koleksi",
-    bookmarksEmptyDesc: "Jelajahi Al-Quran dan simpan ayat yang menyentuh hati Anda.",
-    bookmarksStartReading: "Mulai Membaca",
-    bookmarksLastReadLabel: "Terakhir Dibaca",
-    bookmarksSetLastRead: "Jadikan Terakhir Baca",
-    bookmarksCurrentlyReading: "Sedang Dibaca",
-    bookmarksDelete: "Hapus",
-    bookmarksDeleteConfirm: "Hapus tanda baca ini?",
-    bookmarksMarkedAsLastRead: "Ditandai sebagai Terakhir Baca 📖",
-    bookmarksEditTitle: "Edit Bookmark",
-    bookmarksNotePlaceholder: "Tulis catatan Anda di sini...",
-    bookmarksMarkAsLastRead: "Tandai Terakhir Dibaca",
-    bookmarksUpdateProgress: "Update progress membaca di halaman beranda",
-    bookmarksSave: "Simpan Perubahan",
-    quranSearchPlaceholder: "Cari Surat (e.g. Al-Kahf) atau Arti...",
-    quranSurahList: "Surat-Surat",
-    quranSurahCount: "SURAT",
-    quranVerseCount: "Ayat",
-    quranMakkah: "Mekah",
-    quranMadinah: "Madinah",
-    quranLoadError: "Gagal memuat data Al-Quran",
-    quranLoadErrorDesc: "Mohon periksa koneksi internet Anda dan coba lagi.",
-
-    // Quran Reader Settings
-    quranPlaySurah: "Putar Surat",
-    quranJumpToVerse: "Loncat ke Ayat",
-    quranJumpToVerseTitle: "Loncat ke Ayat",
-    quranGoToVerse: "Pergi ke Ayat",
-    quranSettingsTitle: "Pengaturan Tampilan",
-    quranModeRead: "Mode Baca",
-    quranModeList: "List",
-    quranModeMushaf: "Mushaf",
-    quranScriptType: "Jenis Teks Arab",
-    quranScriptStandard: "Standar Indonesia",
-    quranScriptTajweed: "Tajweed Berwarna",
-    quranOtherDisplay: "Tampilan Lainnya",
-    quranTransliteration: "Latin / Transliterasi",
-    quranFontSize: "Ukuran Huruf Arab",
-    quranVersesPerPage: "Ayat per Halaman",
-    quranSelectQari: "Pilih Qari",
-    quranPrevious: "Sebelumnya",
-    quranNext: "Selanjutnya",
-    quranPage: "Halaman",
-    quranOf: "/",
-
-    // Prayer Names
-    prayerImsak: "Imsak",
-    prayerFajr: "Subuh",
-    prayerSunrise: "Terbit",
-    prayerDhuhr: "Dzuhur",
-    prayerAsr: "Ashar",
-    prayerMaghrib: "Maghrib",
-    prayerIsha: "Isya",
-
-    // Quote of the Day
-    quoteText1: "Maka sesungguhnya bersama kesulitan ada kemudahan.",
-    quoteSource1: "QS. Al-Insyirah: 5",
-    quoteText2: "Dan berikanlah berita gembira kepada orang-orang yang sabar.",
-    quoteSource2: "QS. Al-Baqarah: 155",
-    quoteText3: "Allah tidak membebani seseorang melainkan sesuai dengan kesanggupannya.",
-    quoteSource3: "QS. Al-Baqarah: 286",
-    quoteText4: "Barangsiapa bertakwa kepada Allah niscaya Dia akan mengadakan baginya jalan keluar.",
-    quoteSource4: "QS. At-Talaq: 2",
-    quoteText5: "Dan Dia memberinya rezeki dari arah yang tiada disangka-sangkanya.",
-    quoteSource5: "QS. At-Talaq: 3",
-    quoteText6: "Ingatlah, hanya dengan mengingati Allah-lah hati menjadi tenteram.",
-    quoteSource6: "QS. Ar-Ra'd: 28",
-    quoteText7: "Sebaik-baik manusia adalah yang paling bermanfaat bagi manusia.",
-    quoteSource7: "HR. Ahmad",
-    quoteText8: "Senyummu di hadapan saudaramu adalah sedekah.",
-    quoteSource8: "HR. Tirmidzi",
-    quoteText9: "Janganlah kamu marah, maka bagimu surga.",
-    quoteSource9: "HR. Thabrani",
-    quoteText10: "Sholat itu adalah tiang agama.",
-    quoteSource10: "HR. Tirmidzi",
-
-    // Stats Page
-    statsPageTitle: "Statistik Ibadah",
-    statsPageSubtitle: "Pantau perkembangan spiritualmu",
-    statsWeekly: "Mingguan",
-    statsMonthly: "Bulanan",
-    statsTotalXP: "Total XP",
-    statsComingSoonFeature: "Fitur Sedang Dimatangkan ✨",
-    statsComingSoonMessage: "Kami sedang menyempurnakan algoritma statistik agar jurnal ibadah Anda lebih akurat dan bermakna. Sabar ya, Kak!",
-    statsBackHome: "Kembali ke Beranda",
-
-    // Nawaetu AI Page
-    tanyaTitle: "Nawaetu AI",
-    tanyaSubtitle: "Asisten Muslim Digital",
-    tanyaGreetingMorning: "Assalamu'alaikum kak! 🙏 Semoga hari ini penuh berkah. Ada yang ingin ditanyakan?",
-    tanyaGreetingDay: "Assalamu'alaikum kak! ☀️ Selamat pagi! Aku siap menemani diskusi spiritualmu.",
-    tanyaGreetingNoon: "Assalamu'alaikum kak! 🌤️ Siang ini ada yang mau diceritakan?",
-    tanyaGreetingAfternoon: "Assalamu'alaikum kak! 🌅 Sore yang indah. Yuk ngobrol seputar ibadah.",
-    tanyaGreetingEvening: "Assalamu'alaikum kak! 🌙 Selamat malam. Mau curhat atau tanya hukum islam?",
-    tanyaPlaceholder: "Tanya sesuatu...",
-    tanyaSend: "Kirim",
-    tanyaDailyLimit: "Batas Pertanyaan",
-    tanyaLimitReached: "Kamu sudah mencapai batas 5 pertanyaan hari ini.",
-    tanyaLimitReset: "Limit akan reset besok jam 00:00",
-    tanyaUnlockPremium: "Unlimited untuk Muhsinin",
-    tanyaQuickPrompts: "Prompt Cepat",
-    tanyaPrompt1: "Saya merasa malas sholat...",
-    tanyaPrompt2: "Sedang banyak pikiran",
-    tanyaPrompt3: "Tips khusyuk?",
-    tanyaPrompt4: "Apa keutamaan subuh?",
-
-    // Toast Messages
-    toastMissionComplete: "Misi Selesai!",
-    toastMissionMsg1: "Masya Allah! Istiqamah terus ya ✨",
-    toastMissionMsg2: "Alhamdulillah, satu kebaikan tercatat 🤲",
-    toastMissionMsg3: "Barakallah! Semangat ibadahnya 🔥",
-    toastMissionMsg4: "Keren! Niat lurus, amal mulus 💎",
-
-    // Onboarding
-    onboardingSkip: "Lewati Intro",
-    onboardingBack: "Kembali",
-    onboardingNext: "Lanjut",
-    onboardingFinish: "Selesai & Mulai 🚀",
-    onboardingNameTitle: "Siapa namamu?",
-    onboardingNameDesc: "Agar Nawaetu AI bisa menyapamu dengan akrab.",
-    onboardingNamePlaceholder: "Nama Panggilan",
-    onboardingDefaultName: "Sobat Nawaetu",
-    onboardingGenderTitle: "Apa Jenis Kelaminmu?",
-    onboardingGenderDesc: "Untuk menyesuaikan fiqih ibadah & tema.",
-    onboardingMaleLabel: "Laki-laki",
-    onboardingMaleSub: "Tema Biru, Fiqih Pria",
-    onboardingFemaleLabel: "Perempuan",
-    onboardingFemaleSub: "Tema Pink, Fiqih Wanita",
-    onboardingArchetypeTitle: "Tipe Pejuang Ibadah?",
-    onboardingArchetypeDesc: "Menentukan target misi harianmu.",
-    onboardingArchetypeBeginnerLabel: "Beginner",
-    onboardingArchetypeBeginnerDesc: "Fokus Ibadah Wajib",
-    onboardingArchetypeStriverLabel: "Striver",
-    onboardingArchetypeStriverDesc: "Wajib + Sunnah Rutin",
-    onboardingArchetypeDedicatedLabel: "Dedicated",
-    onboardingArchetypeDedicatedDesc: "Target Ibadah Tinggi",
-
-    // Onboarding Slides
-    onboardingSlide1Title: "Tanpa Install, Hemat Memori",
-    onboardingSlide1Desc: "Tidak perlu download dari PlayStore/AppStore. Cukup akses web ini, ringan dan cepat.",
-    onboardingSlide1Hint: "Hint: Tambahkan ke Home Screen untuk akses instan.",
-    onboardingSlide2Title: "Al-Qur'an & Mentor AI Terpercaya",
-    onboardingSlide2Desc: "Baca Al-Qur'an, tafsir, dan tanya jawab agama. Semua konten AI dipastikan bersumber dari Al-Qur'an & Hadits Shahih.",
-    onboardingSlide2Hint: "Tenang, sumbernya valid & dapat dipertanggungjawabkan.",
-    onboardingSlide3Title: "Gamifikasi Personal",
-    onboardingSlide3Desc: "Misi harian & target XP akan disesuaikan dengan profil Anda (Gender & Tipe Pejuang) agar lebih relevan.",
-    onboardingSlide3Hint: "Ibadah jadi lebih semangat dengan target yang pas.",
-    onboardingSlide4Title: "Tanpa Iklan & Privasi Terjaga",
-    onboardingSlide4Desc: "Fokus ibadah tanpa gangguan iklan. Data tersimpan lokal di HP Anda, tidak kami jual.",
-    onboardingSlide4Hint: "PENTING: Jangan 'Clear Cache' agar data tidak hilang.",
-  },
-  en: {
-    // Header
-    title: "Settings",
-
-    // Stats Section
-    statsLabel: "Worship Statistics",
-    statsComingSoon: "Coming Soon",
-    statsDescription: "Perfect your intention, monitor results later ✨",
-
-    // Location Section
-    locationLabel: "Location",
-    locationUpdating: "Updating...",
-    locationDetecting: "Detecting...",
-
-    // Calculation Method
-    methodLabel: "Method",
-    methodPlaceholder: "Select method",
-
-    // Prayer Notifications
-    notificationTitle: "Prayer Notifications",
-    notificationButton: "Enable Prayer Notifications",
-    notificationPermissionTitle: "Enable Prayer Notifications",
-    notificationPermissionDesc: "To receive accurate prayer time reminders, please allow notification access on your device.",
-    notificationPermissionButton: "Allow Notification Access",
-    notificationDenied: "Notification permission denied. Please enable it in your browser settings.",
-    prayerNotifications: "Prayer Time Notifications",
-    notificationsEnabled: "Notifications enabled",
-    notificationsDisabled: "Notifications disabled",
-    selectPrayerTimes: "Select Prayer Times",
-    notificationLocationNote: "Notifications will be sent based on your saved location in the app.",
-
-    // Prayer Names
-    fajr: "Fajr",
-    dhuhr: "Dhuhr",
-    asr: "Asr",
-    maghrib: "Maghrib",
-    isha: "Isha",
-
-    // Theme
-    themeTitle: "App Appearance",
-
-    // Audio Settings
-    audioTitle: "Audio Settings",
-    muadzinLabel: "Adhan Voice",
-
-    // Language Settings
-    languageTitle: "Language",
-    languageDescription: "The app language will change according to your selection.",
-
-    // Support Card
-    supportTitle: "Support Nawaetu",
-    supportPremiumText: "Thank you for being a Muhsinin! Your donations support server and development.",
-    supportText: "Help us keep the app free and ad-free forever.",
-    infaqButtonPremium: "Donate Again",
-    infaqButton: "Donate",
-
-    // About Section
-    aboutAppName: "Nawaetu",
-    aboutTagline: "#NiatAjaDulu",
-    aboutDescription: "Your digital worship companion for consistent habits based on Quran & Sunnah.",
-    aboutHashtag: "#PerfectYourIntention",
-    aboutVersion: `v${APP_CONFIG.version} • ${APP_CONFIG.status}`,
-    aboutTapForInfo: "Tap for more info",
-    aboutWhatsNew: "What's New",
-    aboutUpdate1: "Mentor AI Chat History (Multi-session support)",
-    aboutUpdate2: "Smart Rate Limiting (Free 5x/day, Muhsinin 25x/day)",
-    aboutUpdate3: "Streamlined AI Responses (Direct & Concise)",
-    aboutUpdate4: "UI & Dark Mode Improvements",
-    aboutUpdate5: "Performance optimizations & bug fixes",
-    aboutFeature1: "Quran",
-    aboutFeature2: "Prayer Times",
-    aboutFeature3: "Mentor AI",
-    aboutFeature4: "Multi-Language",
-    aboutWhatIsTitle: "What is Nawaetu?",
-    aboutApproachTitle: "Our Approach",
-    aboutApproachDesc: "We blend worship guidance with gamification (Missions & Streak) and AI to help you build steadfast habits.",
-    aboutLastUpdate: `Updated: ${APP_CONFIG.lastUpdated}`,
-
-    // Profile Dialog
-    profileGenderLabel: "Gender",
-    profileMale: "Male",
-    profileFemale: "Female",
-    profileGenderHint: "Select to customize missions and features",
-    profileArchetypeLabel: "Worship Focus (Warrior Type)",
-    profileArchetypeBeginner: "Beginner",
-    profileArchetypeMobilizer: "Mobilizer",
-    profileArchetypeWarrior: "Warrior",
-    profileProgressLabel: "Level Progress",
-    profileSupport: "Support",
-    profileAddDonation: "Add Donation",
-    profileTitleUnlock: "Title Unlocked",
-    profileTitleLocked: "Locked (Need Lv.{level})",
-    profilePhilosophy: "Title Philosophy",
-    profileRewards: "Privilege & Rewards",
-    profileUseTitle: "Equip This Title",
-    profileCurrentTitle: "Currently Using",
-    profileMinimalLevel: "Minimal Level {level}",
-    profileSave: "Save Profile",
-    profileTitleLabel: "Worship Title",
-    profileNameLabel: "Full Name",
-    profileAppSettings: "App Settings",
-    profileSyncData: "Sync Data",
-    profileSyncing: "Syncing...",
-    profileShareApp: "Share App",
-    profileLogout: "Logout",
-    profileLogoutConfirm: "Are you sure you want to logout?",
-    profileLogoutConfirmYes: "Yes, Logout",
-    profileLogoutConfirmNo: "Cancel",
-    profileGuestMode: "Guest Mode",
-    profileFooter: "Thank you for using Nawaetu",
-    profileAuthTitle: "Save Your Worship Progress",
-    profileAuthDesc: "Login to save your streak data, bookmarks, and donor status permanently in the cloud.",
-    profileAuthButton: "Login with Google",
-    profileRoleGuest: "Loyal User",
-    profileRolePremium: "Nawaetu Muhsinin",
-    profileStreakLabel: "Streak Days",
-    profileTotalDays: "Total Days",
-
-    // Streak & Progress
-    profileNeedXP: "Need {xp} more XP to reach Level {level}",
-    profileStreak: "Streak",
-    profileRecord: "Record",
-    profileDays: "Days",
-    profileStreakMaintain: "Maintain your commitment every day!",
-    profileBestAchievement: "Your best achievement.",
-    profileTitleCollection: "Title Collection",
-
-    // XP Earning Methods Section
-    xpSectionTitle: "How to Earn XP",
-    xpMethodReadQuran: "Read Quran",
-    xpReadQuranReward: "+5 XP/verse",
-    xpMethodTasbih: "Tasbih",
-    xpTasbihReward: "+50 XP/session",
-    xpMethodDailyMission: "Daily Mission",
-    xpDailyMissionReward: "+25-100 XP/mission",
-    xpMethodStatistics: "Statistics",
-    xpStatisticsHint: "View Progress",
-
-    // Navigation Menu
-    navHome: "Home",
-    navQuran: "Quran",
-    navTasbih: "Tasbih",
-    navQibla: "Qibla",
-    navSettings: "Settings",
-
-    // Qibla Page
-    qiblaFinding: "BREATHE, FIND QIBLA",
-    qiblaAligned: "CALM HEART",
-    qiblaAlignedBadge: "QIBLA SET",
-    qiblaPermissionTitle: "Compass Permission",
-    qiblaPermissionDesc: "Nawaetu needs access to your device's motion sensors to determine the Qibla direction with precision.",
-    qiblaPermissionButton: "Enable Sensor",
-    qiblaLocationError: "Location permission is required to calculate the Qibla direction.",
-    qiblaGeoError: "Geolocation is not supported in this browser.",
-    qiblaCompassDenied: "Compass permission denied.",
-    qiblaCompassFailed: "Failed to request compass permission.",
-    qiblaDistance: "Distance to Ka'bah: {distance} km",
-    qiblaSearching: "Finding location...",
-    qiblaDalilTitle: "Qibla References",
-    qiblaDalilQuranRef: "Quran 2:144",
-    qiblaDalilQuranText: "Turn toward Al-Masjid Al-Haram.",
-    qiblaDalilContextTitle: "Context:",
-    qiblaDalilContextText: "Qibla shifted from Jerusalem to the Ka'bah.",
-
-    // Tasbih Page
-    tasbihTitle: "Digital Tasbih",
-    tasbihSubtitle: "Tranquil Remembrance",
-    tasbihSelectZikir: "Select Zikir",
-    tasbihTarget: "Target",
-    tasbihDaily: "Daily",
-    tasbihStreak: "Streak",
-    tasbihReset: "Reset",
-    tasbihResetConfirm: "Reset the count?",
-    tasbihTap: "Tap",
-    tasbihDays: "Days",
-    tasbihListTitle: "Recitations",
-    tasbihVibrate: "Vibrate",
-    tasbihSound: "Sound",
-    tasbihDual: "Dual",
-    tasbihMute: "Mute",
-    tasbihFreeMode: "Free Mode",
-    tasbihCounterLabel: "Counter",
-    tasbihComplete: "Alhamdulillah!",
-    tasbihCompleteMessage: "Today's zikir target achieved. May it be blessed!",
-    tasbihContinue: "Continue Zikir",
-    tasbihBack: "Back",
-    tasbihNextUp: "Next up",
-    tasbihRepeat: "Repeat",
-    tasbihRepeatReading: "Repeat Recitation",
-    tasbihPresetTasbihLabel: "Tasbih",
-    tasbihPresetTasbihLatin: "Subhanallah",
-    tasbihPresetTasbihTadabbur: "Glory be to Allah, free from all shortcomings.",
-    tasbihPresetTahmidLabel: "Tahmid",
-    tasbihPresetTahmidLatin: "Alhamdulillah",
-    tasbihPresetTahmidTadabbur: "All praise is for Allah for every blessing.",
-    tasbihPresetTakbirLabel: "Takbir",
-    tasbihPresetTakbirLatin: "Allahu Akbar",
-    tasbihPresetTakbirTadabbur: "Allah is the Greatest.",
-    tasbihPresetIstighfarLabel: "Istighfar",
-    tasbihPresetIstighfarLatin: "Astaghfirullah",
-    tasbihPresetIstighfarTadabbur: "I seek Allah's forgiveness.",
-    tasbihPresetSholawatJibrilLabel: "Salawat Jibril",
-    tasbihPresetSholawatJibrilLatin: "Shallallahu 'ala Muhammad",
-    tasbihPresetSholawatJibrilTadabbur: "May Allah send blessings upon Prophet Muhammad.",
-
-    // Quran Page
-    quranTitle: "Digital Qur'an",
-    quranSubtitle: "More than just reading",
-    quranChooseSurah: "Choose Surah",
-    quranBookmark: "Bookmark",
-    quranShare: "Share",
-    quranCopy: "Copy",
-    quranTafsir: "Tafsir",
-    quranAudio: "Audio",
-
-    // Home Page
-    homeGreetingMorning: "Good Morning",
-    homeGreetingNoon: "Good Afternoon",
-    homeGreetingAfternoon: "Good Afternoon",
-    homeGreetingEvening: "Good Evening",
-    homeGreetingWelcome: "Welcome",
-    homeLocation: "Location",
-    homeLoading: "Loading...",
-
-    // Home Widgets
-    widgetLastRead: "Last Read",
-    widgetContinueReading: "Continue Reading",
-    widgetMissions: "Today's Missions",
-    widgetMissionsToday: "Today's Missions",
-    widgetNoMissions: "No missions today",
-    widgetPrayerTimes: "Prayer Times",
-    widgetNextPrayer: "Next Prayer",
-    widgetQuoteOfDay: "Quote of the Day",
-    widgetRamadhan: "Ramadhan",
-    homeAiTitle: "Ask Nawaetu",
-    homeAiSubtitle: "Muslim AI Assistant",
-    homeNextLabel: "Next",
-    homePreparation: "Preparation",
-    homeTowards: "Towards {prayer}",
-    homeAdzanNow: "Adhan {prayer}",
-    homeAdzanMessage: "May your prayer be accepted",
-    homePrayerTimesToday: "Today's Prayer Times",
-    homeFindMosque: "Find Mosque",
-    homeFindMosqueAria: "Find Nearby Mosque via Google Maps",
-
-    // Mosque Finder Modal
-    mosqueFinderTitle: "Where to Pray?",
-    mosqueFinderDesc: "Choose the type of prayer place you need:",
-    mosqueFinderOptionNearby: "Nearest Mosque",
-    mosqueFinderOptionMusholla: "Musholla (Small)",
-    mosqueFinderOptionGrand: "Grand Mosque",
-    mosqueFinderOptionRestArea: "Rest Area",
-    mosqueFinderButton: "Open Google Maps",
-    homeLocationRequiredTitle: "Location Permission Required",
-    homeSetLocationNow: "Set Location",
-    homeLocationDefaultTitle: "Default Location (Jakarta)",
-    homeLocationDefaultDesc: "Current prayer times are based on Jakarta. Tap the location icon above to update to your location.",
-    homeLocationRequiredDesc: "Nawaetu needs your location to calculate accurate prayer times and Qibla direction.",
-    homeEnableLocation: "Enable Location",
-    homeLocationHint: "Your browser will ask for permission after clicking.",
-    homeLastReadStartTitle: "Start Recitation",
-    homeLastReadStartSubtitle: "Read the Quran today",
-    homeLastReadLabel: "Last Read",
-    homeMissionFocusTitle: "Daily Focus",
-    homeMissionDailyTarget: "Your main focus today",
-    homeMissionCompleted: "Completed",
-    homeMissionViewAll: "View All Missions",
-    homeMissionSelectGenderHint: "💡 Select your gender in Profile for more personal missions",
-    homeMissionSpecial: "Special",
-    homeMissionLocked: "Locked",
-    homeMissionLate: "Missed",
-    homeMissionEarly: "Early",
-    homeMissionEarlyTitle: "Virtue of Early Prayer",
-    homeMissionEarlyQuote: "\"The best deed is prayer at its earliest time.\" (HR. Tirmidhi)",
-    homeMissionLateTitle: "Time Almost Up! ({minutes} mins)",
-    homeMissionLateQuote: "\"Do not be negligent about prayer.\" (QS. Al-Ma'un: 5)",
-    homeMissionListTitle: "🎯 Full Mission List",
-    homeMissionLatePrayerTitle: "Missed Prayer (Qadha)",
-    homeMissionEarlyPrayerTitle: "Virtue of Early Prayer",
-    homeMissionLockedFallback: "Locked",
-    homeMissionCompleteLate: "Log Practice (Missed)",
-    homeMissionComplete: "Complete Mission",
-    homeMissionCompletedLabel: "Mission Completed",
-    homeMissionUndoConfirm: "Yes, Undo",
-    homeMissionUndoCancel: "Cancel",
-    homeMissionUndoPrompt: "Undo Completion?",
-    homeMissionLateWarningQuote: "\"So woe to those who pray, who are heedless of their prayer.\" (QS. Al-Ma'un: 4-5)",
-    homeMissionLateWarningDesc: "Complete your prayer even if the time has passed. Better late than never.",
-    homeMissionLateNotice: "The prime time for this mission has passed, but you can still validate if you've done it.",
-    homeMissionEarlyPraiseQuote: "\"The most beloved deed to Allah is prayer at its appointed time.\" (HR. Bukhari & Muslim)",
-    ramadhanHeading: "Towards Ramadhan",
-    ramadhanDaysLeft: "Days Left",
-    ramadhanReady: "Ramadhan Ready! 🌙",
-    ramadhanLevel75: "Consistent Warrior",
-    ramadhanLevel50: "Ready to Step",
-    ramadhanLevel25: "Intention Set",
-    ramadhanLevel0: "Ramadhan Newbie",
-    ramadhanPreparationLabel: "Preparation",
-    ramadhanInfoTitle: "Ramadhan Preparation",
-    ramadhanInfoDesc: "The \"Preparation\" percentage is calculated from completing special missions in Sya'ban (like Sunnah fasting, reading Quran, etc).",
-    ramadhanImproveTitle: "🎯 How to Improve:",
-    ramadhanImproveItem1: "Complete Sya'ban Sunnah fasting missions",
-    ramadhanImproveItem2: "Do Fajr charity",
-    ramadhanImproveItem3: "Start a regular Quran reading habit",
-    ramadhanUnderstand: "Got it",
-    streakNextMilestone: "{days} days to {label}",
-    streakLongest: "Longest streak!",
-    streakDaysLabel: "days",
-
-    // PWA Install Prompt
-    pwaInstallTitle: "Install Nawaetu",
-    pwaInstallDesc: "Faster access, save data, and full-screen features without downloading from the Store.",
-    pwaInstallIosStep1: "Tap the Share button",
-    pwaInstallIosStep2: "Select Add to Home Screen",
-    pwaInstallButton: "Install now",
-    pwaUpdateAvailableTitle: "Update Available",
-    pwaUpdateAvailableDesc: "A new version of Nawaetu is available.",
-    pwaUpdateAction: "Update Now",
-
-
-    // Mission List Modal Tabs
-    missionTabAll: "All",
-    missionTabDaily: "📋 Daily",
-    missionTabWeekly: "📅 Weekly",
-    missionTabRamadhan: "🌙 Ramadhan",
-    missionTabSyaban: "🌙 Sha'ban",
-    missionTabSeasonal: "🌙 Seasonal",
-    missionEmptyCategory: "No missions in this category yet.",
-    missionEmptySeasonalTitle: "No special missions this month.",
-    missionEmptySeasonalDesc: "Stay tuned for special Ramadhan missions soon!",
-
-    // Mission Completion Options
-    missionCompletionAlone: "Pray Alone",
-    missionCompletionCongregation: "Congregational at Mosque",
-
-    // Mission Hukum Labels
-    hukumWajib: "Obligatory",
-    hukumSunnah: "Recommended",
-    hukumMubah: "Permissible",
-    hukumMakruh: "Disliked",
-    hukumHaram: "Forbidden",
-
-    // Quran Browser
-    quranLastRead: "Last Read",
-    quranVerse: "Verse",
-    quranNoHistory: "No reading history",
-    quranStartReading: "Start reading now.",
-    quranBookmarks: "Bookmarks",
-    quranVersesSaved: "verses saved",
-
-    // Bookmarks Page
-    bookmarksTitle: "Bookmark Collection",
-    bookmarksSubtitle: "Your notes and selected verses",
-    bookmarksEmptyTitle: "No bookmarks yet",
-    bookmarksEmptyDesc: "Explore the Quran and save verses that touch your heart.",
-    bookmarksStartReading: "Start Reading",
-    bookmarksLastReadLabel: "Last Read",
-    bookmarksSetLastRead: "Set as Last Read",
-    bookmarksCurrentlyReading: "Currently Reading",
-    bookmarksDelete: "Delete",
-    bookmarksDeleteConfirm: "Delete this bookmark?",
-    bookmarksMarkedAsLastRead: "Marked as Last Read 📖",
-    bookmarksEditTitle: "Edit Bookmark",
-    bookmarksNotePlaceholder: "Write your notes here...",
-    bookmarksMarkAsLastRead: "Mark as Last Read",
-    bookmarksUpdateProgress: "Update reading progress on home page",
-    bookmarksSave: "Save Changes",
-    quranSearchPlaceholder: "Search Surah (e.g. Al-Kahf) or Meaning...",
-    quranSurahList: "Surahs",
-    quranSurahCount: "SURAHS",
-    quranVerseCount: "Verses",
-    quranMakkah: "Makkah",
-    quranMadinah: "Madinah",
-    quranLoadError: "Failed to load Quran data",
-    quranLoadErrorDesc: "Please check your internet connection and try again.",
-
-    // Quran Reader Settings
-    quranPlaySurah: "Play Surah",
-    quranJumpToVerse: "Jump to Verse",
-    quranJumpToVerseTitle: "Jump to Verse",
-    quranGoToVerse: "Go to Verse",
-    quranSettingsTitle: "Display Settings",
-    quranModeRead: "Reading Mode",
-    quranModeList: "List",
-    quranModeMushaf: "Mushaf",
-    quranScriptType: "Arabic Script Type",
-    quranScriptStandard: "Standard Indonesian",
-    quranScriptTajweed: "Colored Tajweed",
-    quranOtherDisplay: "Other Display",
-    quranTransliteration: "Latin / Transliteration",
-    quranFontSize: "Arabic Font Size",
-    quranVersesPerPage: "Verses per Page",
-    quranSelectQari: "Select Qari",
-    quranPrevious: "Previous",
-    quranNext: "Next",
-    quranPage: "Page",
-    quranOf: "of",
-
-    // Prayer Names
-    prayerImsak: "Imsak",
-    prayerFajr: "Fajr",
-    prayerSunrise: "Sunrise",
-    prayerDhuhr: "Dhuhr",
-    prayerAsr: "Asr",
-    prayerMaghrib: "Maghrib",
-    prayerIsha: "Isha",
-
-    // Quote of the Day
-    quoteText1: "Indeed, with hardship comes ease.",
-    quoteSource1: "QS. Al-Insyirah: 5",
-    quoteText2: "And give good tidings to the patient.",
-    quoteSource2: "QS. Al-Baqarah: 155",
-    quoteText3: "Allah does not burden a soul beyond that it can bear.",
-    quoteSource3: "QS. Al-Baqarah: 286",
-    quoteText4: "Whoever fears Allah, He will make for him a way out.",
-    quoteSource4: "QS. At-Talaq: 2",
-    quoteText5: "And He provides for him from sources he never could imagine.",
-    quoteSource5: "QS. At-Talaq: 3",
-    quoteText6: "Verily, in the remembrance of Allah do hearts find rest.",
-    quoteSource6: "QS. Ar-Ra'd: 28",
-    quoteText7: "The best of people are those who are most beneficial to others.",
-    quoteSource7: "HR. Ahmad",
-    quoteText8: "Your smile in the face of your brother is charity.",
-    quoteSource8: "HR. Tirmidzi",
-    quoteText9: "Do not be angry, and Paradise will be yours.",
-    quoteSource9: "HR. Thabrani",
-    quoteText10: "Prayer is the pillar of religion.",
-    quoteSource10: "HR. Tirmidzi",
-
-    // Stats Page
-    statsPageTitle: "Worship Statistics",
-    statsPageSubtitle: "Monitor your spiritual progress",
-    statsWeekly: "Weekly",
-    statsMonthly: "Monthly",
-    statsTotalXP: "Total XP",
-    statsComingSoonFeature: "Feature Coming Soon ✨",
-    statsComingSoonMessage: "We're perfecting our statistics algorithm so your worship journal is more accurate and meaningful. Please be patient!",
-    statsBackHome: "Back to Home",
-
-    // Nawaetu AI Page
-    tanyaTitle: "Nawaetu AI",
-    tanyaSubtitle: "Digital Muslim Assistant",
-    tanyaGreetingMorning: "Assalamu'alaikum! 🙏 Hope your day is full of blessings. What would you like to ask?",
-    tanyaGreetingDay: "Assalamu'alaikum! ☀️ Good morning! I'm ready to accompany your spiritual discussion.",
-    tanyaGreetingNoon: "Assalamu'alaikum! 🌤️ Afternoon, is there anything you'd like to talk about?",
-    tanyaGreetingAfternoon: "Assalamu'alaikum! 🌅 Beautiful afternoon. Let's discuss worship.",
-    tanyaGreetingEvening: "Assalamu'alaikum! 🌙 Good evening. Want to share or ask about Islam?",
-    tanyaPlaceholder: "Ask something...",
-    tanyaSend: "Send",
-    tanyaDailyLimit: "Daily Limit",
-    tanyaLimitReached: "You've reached your 5 questions limit for today.",
-    tanyaLimitReset: "Limit will reset tomorrow at 00:00",
-    tanyaUnlockPremium: "Unlimited for Muhsinin",
-    tanyaQuickPrompts: "Quick Prompts",
-    tanyaPrompt1: "I feel lazy to pray...",
-    tanyaPrompt2: "I'm feeling overwhelmed",
-    tanyaPrompt3: "Tips for khushu?",
-    tanyaPrompt4: "What's the virtue of Fajr?",
-
-    // Toast Messages
-    toastMissionComplete: "Mission Complete!",
-    toastMissionMsg1: "Masya Allah! Keep it consistent ✨",
-    toastMissionMsg2: "Alhamdulillah, one good deed recorded 🤲",
-    toastMissionMsg3: "Barakallah! Keep the spirit high 🔥",
-    toastMissionMsg4: "Awesome! Pure intention, smooth action 💎",
-
-    // Onboarding
-    onboardingSkip: "Skip Intro",
-    onboardingBack: "Back",
-    onboardingNext: "Next",
-    onboardingFinish: "Finish & Start 🚀",
-    onboardingNameTitle: "What is your name?",
-    onboardingNameDesc: "So Nawaetu AI can greet you personally.",
-    onboardingNamePlaceholder: "Nickname",
-    onboardingDefaultName: "Nawaetu Friend",
-    onboardingGenderTitle: "What is your gender?",
-    onboardingGenderDesc: "To customize worship fiqh & theme.",
-    onboardingMaleLabel: "Male",
-    onboardingMaleSub: "Blue Theme, Male Fiqh",
-    onboardingFemaleLabel: "Female",
-    onboardingFemaleSub: "Pink Theme, Female Fiqh",
-    onboardingArchetypeTitle: "Worship Warrior Type?",
-    onboardingArchetypeDesc: "Determines your daily mission targets.",
-    onboardingArchetypeBeginnerLabel: "Beginner",
-    onboardingArchetypeBeginnerDesc: "Focus on Mandatory Worship",
-    onboardingArchetypeStriverLabel: "Striver",
-    onboardingArchetypeStriverDesc: "Mandatory + Regular Sunnah",
-    onboardingArchetypeDedicatedLabel: "Dedicated",
-    onboardingArchetypeDedicatedDesc: "High Worship Targets",
-
-    // Onboarding Slides
-    onboardingSlide1Title: "No Install, Saves Memory",
-    onboardingSlide1Desc: "No need to download from PlayStore/AppStore. Just access this web, light and fast.",
-    onboardingSlide1Hint: "Hint: Add to Home Screen for instant access.",
-    onboardingSlide2Title: "Trusted Quran & AI Mentor",
-    onboardingSlide2Desc: "Read Quran, tafsir, and ask religious questions. All AI content is ensured to come from Quran & Sahih Hadith.",
-    onboardingSlide2Hint: "Rest assured, the source is valid & accountable.",
-    onboardingSlide3Title: "Personal Gamification",
-    onboardingSlide3Desc: "Daily missions & XP targets will be adjusted to your profile (Gender & Warrior Type) for more relevance.",
-    onboardingSlide3Hint: "Worship becomes more exciting with the right targets.",
-    onboardingSlide4Title: "Ad-Free & Private",
-    onboardingSlide4Desc: "Focus on worship without ad distractions. Data is stored locally on your phone, we do not sell it.",
-    onboardingSlide4Hint: "IMPORTANT: Don't 'Clear Cache' so data is not lost.",
-  }
-};
-
-// Helper function to get translation
-export function getSettingsTranslation(locale: string, key: keyof typeof SETTINGS_TRANSLATIONS.id): string {
-  const translations = SETTINGS_TRANSLATIONS[locale as keyof typeof SETTINGS_TRANSLATIONS] || SETTINGS_TRANSLATIONS.id;
-  return translations[key] as string;
-}
