+++ src/data/settings-translations/en.ts
@@ -1,0 +1,506 @@
+import { APP_CONFIG } from "@/config/app-config";
+
+export const SETTINGS_EN = {
+    // Header
+    title: "Settings",
+
+    // Stats Section
+    statsLabel: "Worship Statistics",
+    statsComingSoon: "Coming Soon",
+    statsDescription: "Perfect your intention, monitor results later ✨",
+
+    // Location Section
+    locationLabel: "Location",
+    locationUpdating: "Updating...",
+    locationDetecting: "Detecting...",
+
+    // Calculation Method
+    methodLabel: "Method",
+    methodPlaceholder: "Select method",
+
+    // Prayer Notifications
+    notificationTitle: "Prayer Notifications",
+    notificationButton: "Enable Prayer Notifications",
+    notificationPermissionTitle: "Enable Prayer Notifications",
+    notificationPermissionDesc: "To receive accurate prayer time reminders, please allow notification access on your device.",
+    notificationPermissionButton: "Allow Notification Access",
+    notificationDenied: "Notification permission denied. Please enable it in your browser settings.",
+    prayerNotifications: "Prayer Time Notifications",
+    notificationsEnabled: "Notifications enabled",
+    notificationsDisabled: "Notifications disabled",
+    selectPrayerTimes: "Select Prayer Times",
+    notificationLocationNote: "Notifications will be sent based on your saved location in the app.",
+
+    // Prayer Names
+    fajr: "Fajr",
+    dhuhr: "Dhuhr",
+    asr: "Asr",
+    maghrib: "Maghrib",
+    isha: "Isha",
+
+    // Theme
+    themeTitle: "App Appearance",
+
+    // Audio Settings
+    audioTitle: "Audio Settings",
+    muadzinLabel: "Adhan Voice",
+
+    // Language Settings
+    languageTitle: "Language",
+    languageDescription: "The app language will change according to your selection.",
+
+    // Support Card
+    supportTitle: "Support Nawaetu",
+    supportPremiumText: "Thank you for being a Muhsinin! Your donations support server and development.",
+    supportText: "Help us keep the app free and ad-free forever.",
+    infaqButtonPremium: "Donate Again",
+    infaqButton: "Donate",
+
+    // About Section
+    aboutAppName: "Nawaetu",
+    aboutTagline: "#NiatAjaDulu",
+    aboutDescription: "Your digital worship companion for consistent habits based on Quran & Sunnah.",
+    aboutHashtag: "#PerfectYourIntention",
+    aboutVersion: `v${APP_CONFIG.version} • ${APP_CONFIG.status}`,
+    aboutTapForInfo: "Tap for more info",
+    aboutWhatsNew: "What's New",
+    aboutUpdate1: "Mentor AI Chat History (Multi-session support)",
+    aboutUpdate2: "Smart Rate Limiting (Free 5x/day, Muhsinin 25x/day)",
+    aboutUpdate3: "Streamlined AI Responses (Direct & Concise)",
+    aboutUpdate4: "UI & Dark Mode Improvements",
+    aboutUpdate5: "Performance optimizations & bug fixes",
+    aboutFeature1: "Quran",
+    aboutFeature2: "Prayer Times",
+    aboutFeature3: "Mentor AI",
+    aboutFeature4: "Multi-Language",
+    aboutWhatIsTitle: "What is Nawaetu?",
+    aboutApproachTitle: "Our Approach",
+    aboutApproachDesc: "We blend worship guidance with gamification (Missions & Streak) and AI to help you build steadfast habits.",
+    aboutLastUpdate: `Updated: ${APP_CONFIG.lastUpdated}`,
+
+    // Profile Dialog
+    profileGenderLabel: "Gender",
+    profileMale: "Male",
+    profileFemale: "Female",
+    profileGenderHint: "Select to customize missions and features",
+    profileArchetypeLabel: "Worship Focus (Warrior Type)",
+    profileArchetypeBeginner: "Beginner",
+    profileArchetypeMobilizer: "Mobilizer",
+    profileArchetypeWarrior: "Warrior",
+    profileProgressLabel: "Level Progress",
+    profileSupport: "Support",
+    profileAddDonation: "Add Donation",
+    profileTitleUnlock: "Title Unlocked",
+    profileTitleLocked: "Locked (Need Lv.{level})",
+    profilePhilosophy: "Title Philosophy",
+    profileRewards: "Privilege & Rewards",
+    profileUseTitle: "Equip This Title",
+    profileCurrentTitle: "Currently Using",
+    profileMinimalLevel: "Minimal Level {level}",
+    profileSave: "Save Profile",
+    profileTitleLabel: "Worship Title",
+    profileNameLabel: "Full Name",
+    profileAppSettings: "App Settings",
+    profileSyncData: "Sync Data",
+    profileSyncing: "Syncing...",
+    profileShareApp: "Share App",
+    profileLogout: "Logout",
+    profileLogoutConfirm: "Are you sure you want to logout?",
+    profileLogoutConfirmYes: "Yes, Logout",
+    profileLogoutConfirmNo: "Cancel",
+    profileGuestMode: "Guest Mode",
+    profileFooter: "Thank you for using Nawaetu",
+    profileAuthTitle: "Save Your Worship Progress",
+    profileAuthDesc: "Login to save your streak data, bookmarks, and donor status permanently in the cloud.",
+    profileAuthButton: "Login with Google",
+    profileRoleGuest: "Loyal User",
+    profileRolePremium: "Nawaetu Muhsinin",
+    profileStreakLabel: "Streak Days",
+    profileTotalDays: "Total Days",
+
+    // Streak & Progress
+    profileNeedXP: "Need {xp} more XP to reach Level {level}",
+    profileStreak: "Streak",
+    profileRecord: "Record",
+    profileDays: "Days",
+    profileStreakMaintain: "Maintain your commitment every day!",
+    profileBestAchievement: "Your best achievement.",
+    profileTitleCollection: "Title Collection",
+
+    // XP Earning Methods Section
+    xpSectionTitle: "How to Earn XP",
+    xpMethodReadQuran: "Read Quran",
+    xpReadQuranReward: "+5 XP/verse",
+    xpMethodTasbih: "Tasbih",
+    xpTasbihReward: "+50 XP/session",
+    xpMethodDailyMission: "Daily Mission",
+    xpDailyMissionReward: "+25-100 XP/mission",
+    xpMethodStatistics: "Statistics",
+    xpStatisticsHint: "View Progress",
+
+    // Navigation Menu
+    navHome: "Home",
+    navQuran: "Quran",
+    navTasbih: "Tasbih",
+    navQibla: "Qibla",
+    navSettings: "Settings",
+
+    // Qibla Page
+    qiblaFinding: "BREATHE, FIND QIBLA",
+    qiblaAligned: "CALM HEART",
+    qiblaAlignedBadge: "QIBLA SET",
+    qiblaPermissionTitle: "Compass Permission",
+    qiblaPermissionDesc: "Nawaetu needs access to your device's motion sensors to determine the Qibla direction with precision.",
+    qiblaPermissionButton: "Enable Sensor",
+    qiblaLocationError: "Location permission is required to calculate the Qibla direction.",
+    qiblaGeoError: "Geolocation is not supported in this browser.",
+    qiblaCompassDenied: "Compass permission denied.",
+    qiblaCompassFailed: "Failed to request compass permission.",
+    qiblaDistance: "Distance to Ka'bah: {distance} km",
+    qiblaSearching: "Finding location...",
+    qiblaDalilTitle: "Qibla References",
+    qiblaDalilQuranRef: "Quran 2:144",
+    qiblaDalilQuranText: "Turn toward Al-Masjid Al-Haram.",
+    qiblaDalilContextTitle: "Context:",
+    qiblaDalilContextText: "Qibla shifted from Jerusalem to the Ka'bah.",
+
+    // Tasbih Page
+    tasbihTitle: "Digital Tasbih",
+    tasbihSubtitle: "Tranquil Remembrance",
+    tasbihSelectZikir: "Select Zikir",
+    tasbihTarget: "Target",
+    tasbihDaily: "Daily",
+    tasbihStreak: "Streak",
+    tasbihReset: "Reset",
+    tasbihResetConfirm: "Reset the count?",
+    tasbihTap: "Tap",
+    tasbihDays: "Days",
+    tasbihListTitle: "Recitations",
+    tasbihVibrate: "Vibrate",
+    tasbihSound: "Sound",
+    tasbihDual: "Dual",
+    tasbihMute: "Mute",
+    tasbihFreeMode: "Free Mode",
+    tasbihCounterLabel: "Counter",
+    tasbihComplete: "Alhamdulillah!",
+    tasbihCompleteMessage: "Today's zikir target achieved. May it be blessed!",
+    tasbihContinue: "Continue Zikir",
+    tasbihBack: "Back",
+    tasbihNextUp: "Next up",
+    tasbihRepeat: "Repeat",
+    tasbihRepeatReading: "Repeat Recitation",
+    tasbihPresetTasbihLabel: "Tasbih",
+    tasbihPresetTasbihLatin: "Subhanallah",
+    tasbihPresetTasbihTadabbur: "Glory be to Allah, free from all shortcomings.",
+    tasbihPresetTahmidLabel: "Tahmid",
+    tasbihPresetTahmidLatin: "Alhamdulillah",
+    tasbihPresetTahmidTadabbur: "All praise is for Allah for every blessing.",
+    tasbihPresetTakbirLabel: "Takbir",
+    tasbihPresetTakbirLatin: "Allahu Akbar",
+    tasbihPresetTakbirTadabbur: "Allah is the Greatest.",
+    tasbihPresetIstighfarLabel: "Istighfar",
+    tasbihPresetIstighfarLatin: "Astaghfirullah",
+    tasbihPresetIstighfarTadabbur: "I seek Allah's forgiveness.",
+    tasbihPresetSholawatJibrilLabel: "Salawat Jibril",
+    tasbihPresetSholawatJibrilLatin: "Shallallahu 'ala Muhammad",
+    tasbihPresetSholawatJibrilTadabbur: "May Allah send blessings upon Prophet Muhammad.",
+
+    // Quran Page
+    quranTitle: "Digital Qur'an",
+    quranSubtitle: "More than just reading",
+    quranChooseSurah: "Choose Surah",
+    quranBookmark: "Bookmark",
+    quranShare: "Share",
+    quranCopy: "Copy",
+    quranTafsir: "Tafsir",
+    quranAudio: "Audio",
+
+    // Home Page
+    homeGreetingMorning: "Good Morning",
+    homeGreetingNoon: "Good Afternoon",
+    homeGreetingAfternoon: "Good Afternoon",
+    homeGreetingEvening: "Good Evening",
+    homeGreetingWelcome: "Welcome",
+    homeLocation: "Location",
+    homeLoading: "Loading...",
+
+    // Home Widgets
+    widgetLastRead: "Last Read",
+    widgetContinueReading: "Continue Reading",
+    widgetMissions: "Today's Missions",
+    widgetMissionsToday: "Today's Missions",
+    widgetNoMissions: "No missions today",
+    widgetPrayerTimes: "Prayer Times",
+    widgetNextPrayer: "Next Prayer",
+    widgetQuoteOfDay: "Quote of the Day",
+    widgetRamadhan: "Ramadhan",
+    homeAiTitle: "Ask Nawaetu",
+    homeAiSubtitle: "Muslim AI Assistant",
+    homeNextLabel: "Next",
+    homePreparation: "Preparation",
+    homeTowards: "Towards {prayer}",
+    homeAdzanNow: "Adhan {prayer}",
+    homeAdzanMessage: "May your prayer be accepted",
+    homePrayerTimesToday: "Today's Prayer Times",
+    homeFindMosque: "Find Mosque",
+    homeFindMosqueAria: "Find Nearby Mosque via Google Maps",
+
+    // Mosque Finder Modal
+    mosqueFinderTitle: "Where to Pray?",
+    mosqueFinderDesc: "Choose the type of prayer place you need:",
+    mosqueFinderOptionNearby: "Nearest Mosque",
+    mosqueFinderOptionMusholla: "Musholla (Small)",
+    mosqueFinderOptionGrand: "Grand Mosque",
+    mosqueFinderOptionRestArea: "Rest Area",
+    mosqueFinderButton: "Open Google Maps",
+    homeLocationRequiredTitle: "Location Permission Required",
+    homeSetLocationNow: "Set Location",
+    homeLocationDefaultTitle: "Default Location (Jakarta)",
+    homeLocationDefaultDesc: "Current prayer times are based on Jakarta. Tap the location icon above to update to your location.",
+    homeLocationRequiredDesc: "Nawaetu needs your location to calculate accurate prayer times and Qibla direction.",
+    homeEnableLocation: "Enable Location",
+    homeLocationHint: "Your browser will ask for permission after clicking.",
+    homeLastReadStartTitle: "Start Recitation",
+    homeLastReadStartSubtitle: "Read the Quran today",
+    homeLastReadLabel: "Last Read",
+    homeMissionFocusTitle: "Daily Focus",
+    homeMissionDailyTarget: "Your main focus today",
+    homeMissionCompleted: "Completed",
+    homeMissionViewAll: "View All Missions",
+    homeMissionSelectGenderHint: "💡 Select your gender in Profile for more personal missions",
+    homeMissionSpecial: "Special",
+    homeMissionLocked: "Locked",
+    homeMissionLate: "Missed",
+    homeMissionEarly: "Early",
+    homeMissionEarlyTitle: "Virtue of Early Prayer",
+    homeMissionEarlyQuote: "\"The best deed is prayer at its earliest time.\" (HR. Tirmidhi)",
+    homeMissionLateTitle: "Time Almost Up! ({minutes} mins)",
+    homeMissionLateQuote: "\"Do not be negligent about prayer.\" (QS. Al-Ma'un: 5)",
+    homeMissionListTitle: "🎯 Full Mission List",
+    homeMissionLatePrayerTitle: "Missed Prayer (Qadha)",
+    homeMissionEarlyPrayerTitle: "Virtue of Early Prayer",
+    homeMissionLockedFallback: "Locked",
+    homeMissionCompleteLate: "Log Practice (Missed)",
+    homeMissionComplete: "Complete Mission",
+    homeMissionCompletedLabel: "Mission Completed",
+    homeMissionUndoConfirm: "Yes, Undo",
+    homeMissionUndoCancel: "Cancel",
+    homeMissionUndoPrompt: "Undo Completion?",
+    homeMissionLateWarningQuote: "\"So woe to those who pray, who are heedless of their prayer.\" (QS. Al-Ma'un: 4-5)",
+    homeMissionLateWarningDesc: "Complete your prayer even if the time has passed. Better late than never.",
+    homeMissionLateNotice: "The prime time for this mission has passed, but you can still validate if you've done it.",
+    homeMissionEarlyPraiseQuote: "\"The most beloved deed to Allah is prayer at its appointed time.\" (HR. Bukhari & Muslim)",
+    ramadhanHeading: "Towards Ramadhan",
+    ramadhanDaysLeft: "Days Left",
+    ramadhanReady: "Ramadhan Ready! 🌙",
+    ramadhanLevel75: "Consistent Warrior",
+    ramadhanLevel50: "Ready to Step",
+    ramadhanLevel25: "Intention Set",
+    ramadhanLevel0: "Ramadhan Newbie",
+    ramadhanPreparationLabel: "Preparation",
+    ramadhanInfoTitle: "Ramadhan Preparation",
+    ramadhanInfoDesc: "The \"Preparation\" percentage is calculated from completing special missions in Sya'ban (like Sunnah fasting, reading Quran, etc).",
+    ramadhanImproveTitle: "🎯 How to Improve:",
+    ramadhanImproveItem1: "Complete Sya'ban Sunnah fasting missions",
+    ramadhanImproveItem2: "Do Fajr charity",
+    ramadhanImproveItem3: "Start a regular Quran reading habit",
+    ramadhanUnderstand: "Got it",
+    streakNextMilestone: "{days} days to {label}",
+    streakLongest: "Longest streak!",
+    streakDaysLabel: "days",
+
+    // PWA Install Prompt
+    pwaInstallTitle: "Install Nawaetu",
+    pwaInstallDesc: "Faster access, save data, and full-screen features without downloading from the Store.",
+    pwaInstallIosStep1: "Tap the Share button",
+    pwaInstallIosStep2: "Select Add to Home Screen",
+    pwaInstallButton: "Install now",
+    pwaUpdateAvailableTitle: "Update Available",
+    pwaUpdateAvailableDesc: "A new version of Nawaetu is available.",
+    pwaUpdateAction: "Update Now",
+
+
+    // Mission List Modal Tabs
+    missionTabAll: "All",
+    missionTabDaily: "📋 Daily",
+    missionTabWeekly: "📅 Weekly",
+    missionTabRamadhan: "🌙 Ramadhan",
+    missionTabSyaban: "🌙 Sha'ban",
+    missionTabSeasonal: "🌙 Seasonal",
+    missionEmptyCategory: "No missions in this category yet.",
+    missionEmptySeasonalTitle: "No special missions this month.",
+    missionEmptySeasonalDesc: "Stay tuned for special Ramadhan missions soon!",
+
+    // Mission Completion Options
+    missionCompletionAlone: "Pray Alone",
+    missionCompletionCongregation: "Congregational at Mosque",
+
+    // Mission Hukum Labels
+    hukumWajib: "Obligatory",
+    hukumSunnah: "Recommended",
+    hukumMubah: "Permissible",
+    hukumMakruh: "Disliked",
+    hukumHaram: "Forbidden",
+
+    // Quran Browser
+    quranLastRead: "Last Read",
+    quranVerse: "Verse",
+    quranNoHistory: "No reading history",
+    quranStartReading: "Start reading now.",
+    quranBookmarks: "Bookmarks",
+    quranVersesSaved: "verses saved",
+
+    // Bookmarks Page
+    bookmarksTitle: "Bookmark Collection",
+    bookmarksSubtitle: "Your notes and selected verses",
+    bookmarksEmptyTitle: "No bookmarks yet",
+    bookmarksEmptyDesc: "Explore the Quran and save verses that touch your heart.",
+    bookmarksStartReading: "Start Reading",
+    bookmarksLastReadLabel: "Last Read",
+    bookmarksSetLastRead: "Set as Last Read",
+    bookmarksCurrentlyReading: "Currently Reading",
+    bookmarksDelete: "Delete",
+    bookmarksDeleteConfirm: "Delete this bookmark?",
+    bookmarksMarkedAsLastRead: "Marked as Last Read 📖",
+    bookmarksEditTitle: "Edit Bookmark",
+    bookmarksNotePlaceholder: "Write your notes here...",
+    bookmarksMarkAsLastRead: "Mark as Last Read",
+    bookmarksUpdateProgress: "Update reading progress on home page",
+    bookmarksSave: "Save Changes",
+    quranSearchPlaceholder: "Search Surah (e.g. Al-Kahf) or Meaning...",
+    quranSurahList: "Surahs",
+    quranSurahCount: "SURAHS",
+    quranVerseCount: "Verses",
+    quranMakkah: "Makkah",
+    quranMadinah: "Madinah",
+    quranLoadError: "Failed to load Quran data",
+    quranLoadErrorDesc: "Please check your internet connection and try again.",
+
+    // Quran Reader Settings
+    quranPlaySurah: "Play Surah",
+    quranJumpToVerse: "Jump to Verse",
+    quranJumpToVerseTitle: "Jump to Verse",
+    quranGoToVerse: "Go to Verse",
+    quranSettingsTitle: "Display Settings",
+    quranModeRead: "Reading Mode",
+    quranModeList: "List",
+    quranModeMushaf: "Mushaf",
+    quranScriptType: "Arabic Script Type",
+    quranScriptStandard: "Standard Indonesian",
+    quranScriptTajweed: "Colored Tajweed",
+    quranOtherDisplay: "Other Display",
+    quranTransliteration: "Latin / Transliteration",
+    quranFontSize: "Arabic Font Size",
+    quranVersesPerPage: "Verses per Page",
+    quranSelectQari: "Select Qari",
+    quranPrevious: "Previous",
+    quranNext: "Next",
+    quranPage: "Page",
+    quranOf: "of",
+
+    // Prayer Names
+    prayerImsak: "Imsak",
+    prayerFajr: "Fajr",
+    prayerSunrise: "Sunrise",
+    prayerDhuhr: "Dhuhr",
+    prayerAsr: "Asr",
+    prayerMaghrib: "Maghrib",
+    prayerIsha: "Isha",
+
+    // Quote of the Day
+    quoteText1: "Indeed, with hardship comes ease.",
+    quoteSource1: "QS. Al-Insyirah: 5",
+    quoteText2: "And give good tidings to the patient.",
+    quoteSource2: "QS. Al-Baqarah: 155",
+    quoteText3: "Allah does not burden a soul beyond that it can bear.",
+    quoteSource3: "QS. Al-Baqarah: 286",
+    quoteText4: "Whoever fears Allah, He will make for him a way out.",
+    quoteSource4: "QS. At-Talaq: 2",
+    quoteText5: "And He provides for him from sources he never could imagine.",
+    quoteSource5: "QS. At-Talaq: 3",
+    quoteText6: "Verily, in the remembrance of Allah do hearts find rest.",
+    quoteSource6: "QS. Ar-Ra'd: 28",
+    quoteText7: "The best of people are those who are most beneficial to others.",
+    quoteSource7: "HR. Ahmad",
+    quoteText8: "Your smile in the face of your brother is charity.",
+    quoteSource8: "HR. Tirmidzi",
+    quoteText9: "Do not be angry, and Paradise will be yours.",
+    quoteSource9: "HR. Thabrani",
+    quoteText10: "Prayer is the pillar of religion.",
+    quoteSource10: "HR. Tirmidzi",
+
+    // Stats Page
+    statsPageTitle: "Worship Statistics",
+    statsPageSubtitle: "Monitor your spiritual progress",
+    statsWeekly: "Weekly",
+    statsMonthly: "Monthly",
+    statsTotalXP: "Total XP",
+    statsComingSoonFeature: "Feature Coming Soon ✨",
+    statsComingSoonMessage: "We're perfecting our statistics algorithm so your worship journal is more accurate and meaningful. Please be patient!",
+    statsBackHome: "Back to Home",
+
+    // Nawaetu AI Page
+    tanyaTitle: "Nawaetu AI",
+    tanyaSubtitle: "Digital Muslim Assistant",
+    tanyaGreetingMorning: "Assalamu'alaikum! 🙏 Hope your day is full of blessings. What would you like to ask?",
+    tanyaGreetingDay: "Assalamu'alaikum! ☀️ Good morning! I'm ready to accompany your spiritual discussion.",
+    tanyaGreetingNoon: "Assalamu'alaikum! 🌤️ Afternoon, is there anything you'd like to talk about?",
+    tanyaGreetingAfternoon: "Assalamu'alaikum! 🌅 Beautiful afternoon. Let's discuss worship.",
+    tanyaGreetingEvening: "Assalamu'alaikum! 🌙 Good evening. Want to share or ask about Islam?",
+    tanyaPlaceholder: "Ask something...",
+    tanyaSend: "Send",
+    tanyaDailyLimit: "Daily Limit",
+    tanyaLimitReached: "You've reached your 5 questions limit for today.",
+    tanyaLimitReset: "Limit will reset tomorrow at 00:00",
+    tanyaUnlockPremium: "Unlimited for Muhsinin",
+    tanyaQuickPrompts: "Quick Prompts",
+    tanyaPrompt1: "I feel lazy to pray...",
+    tanyaPrompt2: "I'm feeling overwhelmed",
+    tanyaPrompt3: "Tips for khushu?",
+    tanyaPrompt4: "What's the virtue of Fajr?",
+
+    // Toast Messages
+    toastMissionComplete: "Mission Complete!",
+    toastMissionMsg1: "Masya Allah! Keep it consistent ✨",
+    toastMissionMsg2: "Alhamdulillah, one good deed recorded 🤲",
+    toastMissionMsg3: "Barakallah! Keep the spirit high 🔥",
+    toastMissionMsg4: "Awesome! Pure intention, smooth action 💎",
+
+    // Onboarding
+    onboardingSkip: "Skip Intro",
+    onboardingBack: "Back",
+    onboardingNext: "Next",
+    onboardingFinish: "Finish & Start 🚀",
+    onboardingNameTitle: "What is your name?",
+    onboardingNameDesc: "So Nawaetu AI can greet you personally.",
+    onboardingNamePlaceholder: "Nickname",
+    onboardingDefaultName: "Nawaetu Friend",
+    onboardingGenderTitle: "What is your gender?",
+    onboardingGenderDesc: "To customize worship fiqh & theme.",
+    onboardingMaleLabel: "Male",
+    onboardingMaleSub: "Blue Theme, Male Fiqh",
+    onboardingFemaleLabel: "Female",
+    onboardingFemaleSub: "Pink Theme, Female Fiqh",
+    onboardingArchetypeTitle: "Worship Warrior Type?",
+    onboardingArchetypeDesc: "Determines your daily mission targets.",
+    onboardingArchetypeBeginnerLabel: "Beginner",
+    onboardingArchetypeBeginnerDesc: "Focus on Mandatory Worship",
+    onboardingArchetypeStriverLabel: "Striver",
+    onboardingArchetypeStriverDesc: "Mandatory + Regular Sunnah",
+    onboardingArchetypeDedicatedLabel: "Dedicated",
+    onboardingArchetypeDedicatedDesc: "High Worship Targets",
+
+    // Onboarding Slides
+    onboardingSlide1Title: "No Install, Saves Memory",
+    onboardingSlide1Desc: "No need to download from PlayStore/AppStore. Just access this web, light and fast.",
+    onboardingSlide1Hint: "Hint: Add to Home Screen for instant access.",
+    onboardingSlide2Title: "Trusted Quran & AI Mentor",
+    onboardingSlide2Desc: "Read Quran, tafsir, and ask religious questions. All AI content is ensured to come from Quran & Sahih Hadith.",
+    onboardingSlide2Hint: "Rest assured, the source is valid & accountable.",
+    onboardingSlide3Title: "Personal Gamification",
+    onboardingSlide3Desc: "Daily missions & XP targets will be adjusted to your profile (Gender & Warrior Type) for more relevance.",
+    onboardingSlide3Hint: "Worship becomes more exciting with the right targets.",
+    onboardingSlide4Title: "Ad-Free & Private",
+    onboardingSlide4Desc: "Focus on worship without ad distractions. Data is stored locally on your phone, we do not sell it.",
+    onboardingSlide4Hint: "IMPORTANT: Don't 'Clear Cache' so data is not lost.",
+};
