+++ src/data/settings-translations/id.ts
@@ -1,0 +1,506 @@
+import { APP_CONFIG } from "@/config/app-config";
+
+export const SETTINGS_ID = {
+    // Header
+    title: "Pengaturan",
+
+    // Stats Section
+    statsLabel: "Statistik Ibadah",
+    statsComingSoon: "Segera Hadir",
+    statsDescription: "Matangkan niat, pantau hasil nanti ✨",
+
+    // Location Section
+    locationLabel: "Lokasi",
+    locationUpdating: "Memperbarui...",
+    locationDetecting: "Mendeteksi...",
+
+    // Calculation Method
+    methodLabel: "Metode",
+    methodPlaceholder: "Pilih metode",
+
+    // Prayer Notifications
+    notificationTitle: "Notifikasi Adzan",
+    notificationButton: "Aktifkan Notifikasi Adzan",
+    notificationPermissionTitle: "Aktifkan Notifikasi Sholat",
+    notificationPermissionDesc: "Untuk mendapatkan pengingat waktu sholat yang akurat, mohon izinkan akses notifikasi pada perangkat Anda.",
+    notificationPermissionButton: "Izinkan Akses Notifikasi",
+    notificationDenied: "Izin notifikasi ditolak. Mohon aktifkan di pengaturan browser Anda.",
+    prayerNotifications: "Notifikasi Waktu Sholat",
+    notificationsEnabled: "Notifikasi aktif",
+    notificationsDisabled: "Notifikasi nonaktif",
+    selectPrayerTimes: "Pilih Waktu Sholat",
+    notificationLocationNote: "Notifikasi akan dikirim berdasarkan lokasi yang tersimpan di aplikasi.",
+
+    // Prayer Names
+    fajr: "Subuh",
+    dhuhr: "Dzuhur",
+    asr: "Ashar",
+    maghrib: "Maghrib",
+    isha: "Isya",
+
+    // Theme
+    themeTitle: "Tampilan Aplikasi",
+
+    // Audio Settings
+    audioTitle: "Pengaturan Audio",
+    muadzinLabel: "Suara Adzan",
+
+    // Language Settings
+    languageTitle: "Bahasa",
+    languageDescription: "Bahasa aplikasi akan berubah sesuai pilihan.",
+
+    // Support Card
+    supportTitle: "Dukung Nawaetu",
+    supportPremiumText: "Terima kasih telah menjadi Muhsinin! Hasil infaq digunakan untuk server & pengembangan.",
+    supportText: "Bantu kami menjaga aplikasi tetap gratis dan bebas iklan selamanya.",
+    infaqButtonPremium: "Infaq Lagi",
+    infaqButton: "Infaq",
+
+    // About Section
+    aboutAppName: "Nawaetu",
+    aboutTagline: "Luruskan Niat, Sempurnakan Ibadah",
+    aboutDescription: "Teman ibadah digital yang menjaga setiap langkahmu tetap sesuai tuntunan Al-Qur'an & Sunnah.",
+    aboutHashtag: "#LuruskanNiat",
+    aboutVersion: `v${APP_CONFIG.version} • ${APP_CONFIG.status}`,
+    aboutTapForInfo: "Ketuk untuk info lengkap",
+    aboutWhatsNew: "Apa yang Baru",
+    aboutUpdate1: "Riwayat Chat AI tersimpan antar sesi",
+    aboutUpdate2: "Rate Limit cerdas (Free 5x/hari, Muhsinin 25x/hari)",
+    aboutUpdate3: "Update Respon AI (Lebih ringkas & tanpa basa-basi)",
+    aboutUpdate4: "Perbaikan UI Navigasi & Dark Mode",
+    aboutUpdate5: "Optimasi performa & perbaikan bug",
+    aboutFeature1: "Al-Qur'an",
+    aboutFeature2: "Waktu Sholat",
+    aboutFeature3: "Mentor AI",
+    aboutFeature4: "Multi-Bahasa",
+    aboutWhatIsTitle: "Apa itu Nawaetu?",
+    aboutApproachTitle: "Pendekatan Kami",
+    aboutApproachDesc: "Kami menggabungkan panduan ibadah dengan gamifikasi (Misi & Streak) serta asisten AI untuk membangun kebiasaan yang istiqomah.",
+    aboutLastUpdate: `Diperbarui: ${APP_CONFIG.lastUpdated}`,
+
+    // Profile Dialog
+    profileGenderLabel: "Jenis Kelamin",
+    profileMale: "Laki-laki",
+    profileFemale: "Perempuan",
+    profileGenderHint: "Pilih untuk menyesuaikan misi dan fitur",
+    profileArchetypeLabel: "Fokus Ibadah (Tipe Pejuang)",
+    profileArchetypeBeginner: "Pemula",
+    profileArchetypeMobilizer: "Penggerak",
+    profileArchetypeWarrior: "Mujahid",
+    profileProgressLabel: "Progress Level",
+    profileSupport: "Dukung",
+    profileAddDonation: "Tambah Infaq",
+    profileTitleUnlock: "Gelar Terbuka",
+    profileTitleLocked: "Terkunci (Butuh Lv.{level})",
+    profilePhilosophy: "Filosofi Gelar",
+    profileRewards: "Privilege & Rewards",
+    profileUseTitle: "Pasang Gelar Ini",
+    profileCurrentTitle: "Sedang Digunakan",
+    profileMinimalLevel: "Minimal Level {level}",
+    profileSave: "Simpan Profil",
+    profileTitleLabel: "Gelar Ibadah",
+    profileNameLabel: "Nama Lengkap",
+    profileAppSettings: "Pengaturan Aplikasi",
+    profileSyncData: "Sinkronisasi Data",
+    profileSyncing: "Menyinkronkan...",
+    profileShareApp: "Bagikan Aplikasi",
+    profileLogout: "Keluar",
+    profileLogoutConfirm: "Yakin ingin keluar akun?",
+    profileLogoutConfirmYes: "Ya, Keluar",
+    profileLogoutConfirmNo: "Batal",
+    profileGuestMode: "Mode Guest",
+    profileFooter: "Terima kasih telah menggunakan Nawaetu",
+    profileAuthTitle: "Simpan Progress Ibadahmu",
+    profileAuthDesc: "Login untuk menyimpan data streak, bookmark, dan status donatur secara permanen di cloud.",
+    profileAuthButton: "Login dengan Google",
+    profileRoleGuest: "Pengguna Setia",
+    profileRolePremium: "Muhsinin Nawaetu",
+    profileStreakLabel: "Hari Streak",
+    profileTotalDays: "Total Hari",
+
+    // Streak & Progress
+    profileNeedXP: "Butuh {xp} XP lagi untuk naik ke Level {level}",
+    profileStreak: "Streak",
+    profileRecord: "Rekor",
+    profileDays: "Hari",
+    profileStreakMaintain: "Pertahankan istiqamahmu setiap hari!",
+    profileBestAchievement: "Pencapaian istiqamah terbaikmu.",
+    profileTitleCollection: "Koleksi Gelar",
+
+    // XP Earning Methods Section
+    xpSectionTitle: "Cara Mendapatkan XP",
+    xpMethodReadQuran: "Baca Quran",
+    xpReadQuranReward: "+5 XP/ayat",
+    xpMethodTasbih: "Tasbih",
+    xpTasbihReward: "+50 XP/sesi",
+    xpMethodDailyMission: "Misi Harian",
+    xpDailyMissionReward: "+25-100 XP/misi",
+    xpMethodStatistics: "Statistik",
+    xpStatisticsHint: "Lihat Progress",
+
+    // Navigation Menu
+    navHome: "Beranda",
+    navQuran: "Al-Qur'an",
+    navTasbih: "Tasbih",
+    navQibla: "Kiblat",
+    navSettings: "Atur",
+
+    // Qibla Page
+    qiblaFinding: "TENANG, CARI KIBLAT",
+    qiblaAligned: "HATI TENANG",
+    qiblaAlignedBadge: "KIBLAT TEPAT",
+    qiblaPermissionTitle: "Izin Akses Kompas",
+    qiblaPermissionDesc: "Nawaetu memerlukan akses sensor gerak HP Anda untuk menentukan arah kiblat dengan presisi.",
+    qiblaPermissionButton: "Aktifkan Sensor",
+    qiblaLocationError: "Izin lokasi diperlukan untuk menghitung arah kiblat.",
+    qiblaGeoError: "Geolocation tidak didukung di browser ini.",
+    qiblaCompassDenied: "Izin kompas ditolak.",
+    qiblaCompassFailed: "Gagal meminta izin kompas.",
+    qiblaDistance: "Jarak ke Ka'bah: {distance} km",
+    qiblaSearching: "Mencari lokasi...",
+    qiblaDalilTitle: "Dalil Arah Kiblat",
+    qiblaDalilQuranRef: "QS. Al-Baqarah 2:144",
+    qiblaDalilQuranText: "Hadapkan wajah ke Masjidil Haram.",
+    qiblaDalilContextTitle: "Asbabun nuzul:",
+    qiblaDalilContextText: "Peralihan kiblat dari Baitul Maqdis ke Ka'bah.",
+
+    // Tasbih Page
+    tasbihTitle: "Tasbih Digital",
+    tasbihSubtitle: "Zikir Penenang Hati",
+    tasbihSelectZikir: "Pilih Zikir",
+    tasbihTarget: "Target",
+    tasbihDaily: "Harian",
+    tasbihStreak: "Streak",
+    tasbihReset: "Reset",
+    tasbihResetConfirm: "Reset hitungan?",
+    tasbihTap: "Tap",
+    tasbihDays: "Hari",
+    tasbihListTitle: "Daftar Bacaan",
+    tasbihVibrate: "Getar",
+    tasbihSound: "Suara",
+    tasbihDual: "Dual",
+    tasbihMute: "Mute",
+    tasbihFreeMode: "Mode Bebas",
+    tasbihCounterLabel: "Counter",
+    tasbihComplete: "Alhamdulillah!",
+    tasbihCompleteMessage: "Target zikir hari ini tercapai. Semoga berkah!",
+    tasbihContinue: "Lanjut Zikir",
+    tasbihBack: "Kembali",
+    tasbihNextUp: "Lanjut ke",
+    tasbihRepeat: "Ulangi",
+    tasbihRepeatReading: "Ulangi Bacaan",
+    tasbihPresetTasbihLabel: "Tasbih",
+    tasbihPresetTasbihLatin: "Subhanallah",
+    tasbihPresetTasbihTadabbur: "Maha Suci Allah dari segala kekurangan.",
+    tasbihPresetTahmidLabel: "Tahmid",
+    tasbihPresetTahmidLatin: "Alhamdulillah",
+    tasbihPresetTahmidTadabbur: "Segala puji bagi Allah atas segala nikmat.",
+    tasbihPresetTakbirLabel: "Takbir",
+    tasbihPresetTakbirLatin: "Allahu Akbar",
+    tasbihPresetTakbirTadabbur: "Allah Maha Besar.",
+    tasbihPresetIstighfarLabel: "Istighfar",
+    tasbihPresetIstighfarLatin: "Astaghfirullah",
+    tasbihPresetIstighfarTadabbur: "Aku memohon ampunan kepada Allah SWT.",
+    tasbihPresetSholawatJibrilLabel: "Sholawat Jibril",
+    tasbihPresetSholawatJibrilLatin: "Shallallahu 'ala Muhammad",
+    tasbihPresetSholawatJibrilTadabbur: "Semoga Allah melimpahkan rahmat kepada Nabi Muhammad.",
+
+    // Quran Page
+    quranTitle: "Al-Qur'an Digital",
+    quranSubtitle: "Lebih dari sekedar membaca",
+    quranChooseSurah: "Pilih Surah",
+    quranBookmark: "Bookmark",
+    quranShare: "Bagikan",
+    quranCopy: "Salin",
+    quranTafsir: "Tafsir",
+    quranAudio: "Audio",
+
+    // Home Page
+    homeGreetingMorning: "Selamat Pagi",
+    homeGreetingNoon: "Selamat Siang",
+    homeGreetingAfternoon: "Selamat Sore",
+    homeGreetingEvening: "Selamat Malam",
+    homeGreetingWelcome: "Selamat Datang",
+    homeLocation: "Lokasi",
+    homeLoading: "Loading...",
+
+    // Home Widgets
+    widgetLastRead: "Terakhir Dibaca",
+    widgetContinueReading: "Lanjut Baca",
+    widgetMissions: "Misi Hari Ini",
+    widgetMissionsToday: "Misi Hari Ini",
+    widgetNoMissions: "Belum ada misi hari ini",
+    widgetPrayerTimes: "Jadwal Sholat",
+    widgetNextPrayer: "Sholat Berikutnya",
+    widgetQuoteOfDay: "Quote Hari Ini",
+    widgetRamadhan: "Ramadhan",
+    homeAiTitle: "Tanya Nawaetu",
+    homeAiSubtitle: "Asisten Muslim AI",
+    homeNextLabel: "Selanjutnya",
+    homePreparation: "Persiapan",
+    homeTowards: "Menuju {prayer}",
+    homeAdzanNow: "Adzan {prayer}",
+    homeAdzanMessage: "Selamat menunaikan sholat",
+    homePrayerTimesToday: "Jadwal Sholat Hari Ini",
+    homeFindMosque: "Cari Masjid",
+    homeFindMosqueAria: "Cari Masjid Terdekat via Google Maps",
+
+    // Mosque Finder Modal
+    mosqueFinderTitle: "Mau Sholat di Mana?",
+    mosqueFinderDesc: "Pilih jenis tempat sholat yang kamu butuhkan:",
+    mosqueFinderOptionNearby: "Masjid Terdekat",
+    mosqueFinderOptionMusholla: "Musholla",
+    mosqueFinderOptionGrand: "Masjid Raya",
+    mosqueFinderOptionRestArea: "Rest Area",
+    mosqueFinderButton: "Buka Google Maps",
+    homeLocationRequiredTitle: "Izin Lokasi Diperlukan",
+    homeSetLocationNow: "Set Lokasi",
+    homeLocationDefaultTitle: "Lokasi Default (Jakarta)",
+    homeLocationDefaultDesc: "Jadwal sholat saat ini berdasarkan Jakarta. Ketuk ikon lokasi di atas untuk memperbarui ke lokasi Anda.",
+    homeLocationRequiredDesc: "Nawaetu membutuhkan lokasi Anda untuk menghitung jadwal sholat dan arah kiblat yang akurat.",
+    homeEnableLocation: "Aktifkan Lokasi",
+    homeLocationHint: "Browser akan meminta izin setelah tombol diklik.",
+    homeLastReadStartTitle: "Mulai Tilawah",
+    homeLastReadStartSubtitle: "Baca Al-Quran hari ini",
+    homeLastReadLabel: "Terakhir Baca",
+    homeMissionFocusTitle: "Fokus Harian",
+    homeMissionDailyTarget: "Fokus utama hari ini",
+    homeMissionCompleted: "Selesai",
+    homeMissionViewAll: "Lihat Semua Misi",
+    homeMissionSelectGenderHint: "💡 Pilih jenis kelamin di Profil untuk misi yang lebih personal",
+    homeMissionSpecial: "Special",
+    homeMissionLocked: "Terkunci",
+    homeMissionLate: "Terlewat",
+    homeMissionEarly: "Awal Waktu",
+    homeMissionEarlyTitle: "Keutamaan Awal Waktu",
+    homeMissionEarlyQuote: "\"Amalan terbaik: Shalat di awal waktu.\" (HR. Tirmidzi)",
+    homeMissionLateTitle: "Waktu Hampir Habis! ({minutes} menit)",
+    homeMissionLateQuote: "\"Janganlah kalian lalai terhadap shalat.\" (QS. Al-Ma'un: 5)",
+    homeMissionListTitle: "🎯 Daftar Misi Lengkap",
+    homeMissionLatePrayerTitle: "Sholat Terlewat (Qadha)",
+    homeMissionEarlyPrayerTitle: "Keutamaan Sholat Awal Waktu",
+    homeMissionLockedFallback: "Terkunci",
+    homeMissionCompleteLate: "Catat Amalan (Terlewat)",
+    homeMissionComplete: "Selesaikan Misi",
+    homeMissionCompletedLabel: "Misi Sudah Selesai",
+    homeMissionUndoConfirm: "Ya, Batalkan",
+    homeMissionUndoCancel: "Batal",
+    homeMissionUndoPrompt: "Batalkan Penyelesaian?",
+    homeMissionLateWarningQuote: "\"Maka celakalah bagi orang-orang yang shalat, (yaitu) orang-orang yang lalai dari shalatnya.\" (QS. Al-Ma'un: 4-5)",
+    homeMissionLateWarningDesc: "Segera tunaikan sholatmu meskipun waktunya sudah lewat. Lebih baik terlambat daripada tidak sama sekali.",
+    homeMissionLateNotice: "Waktu utama misi ini sudah lewat, namun Kamu tetap bisa memvalidasi jika sudah mengamalkannya.",
+    homeMissionEarlyPraiseQuote: "\"Amalan yang paling dicintai Allah adalah sholat pada waktunya.\" (HR. Bukhari & Muslim)",
+    ramadhanHeading: "Menuju Ramadhan",
+    ramadhanDaysLeft: "Hari Lagi",
+    ramadhanReady: "Ramadhan Ready! 🌙",
+    ramadhanLevel75: "Pejuang Istiqomah",
+    ramadhanLevel50: "Siap Melangkah",
+    ramadhanLevel25: "Niat Terpasang",
+    ramadhanLevel0: "Newbie Ramadhan",
+    ramadhanPreparationLabel: "Persiapan",
+    ramadhanInfoTitle: "Persiapan Ramadhan",
+    ramadhanInfoDesc: "Persentase \"Persiapan\" dihitung dari penyelesaian misi-misi khusus di bulan Sya'ban (seperti Puasa Sunnah, Baca Quran, dll).",
+    ramadhanImproveTitle: "🎯 Cara Meningkatkan:",
+    ramadhanImproveItem1: "Selesaikan misi Puasa Sunnah Sya'ban",
+    ramadhanImproveItem2: "Lakukan Sedekah Subuh",
+    ramadhanImproveItem3: "Mulai rutinkan baca Al-Quran",
+    ramadhanUnderstand: "Mengerti",
+    streakNextMilestone: "{days} hari lagi ke {label}",
+    streakLongest: "Streak terpanjang!",
+    streakDaysLabel: "hari",
+
+    // PWA Install Prompt
+    pwaInstallTitle: "Install Nawaetu",
+    pwaInstallDesc: "Akses lebih cepat, hemat kuota, dan fitur full screen tanpa perlu download di Store.",
+    pwaInstallIosStep1: "Tap tombol Share",
+    pwaInstallIosStep2: "Pilih Add to Home Screen",
+    pwaInstallButton: "Install sekarang",
+    pwaUpdateAvailableTitle: "Update Tersedia",
+    pwaUpdateAvailableDesc: "Versi baru Nawaetu telah tersedia.",
+    pwaUpdateAction: "Update Sekarang",
+
+
+    // Mission List Modal Tabs
+    missionTabAll: "Semua",
+    missionTabDaily: "📋 Harian",
+    missionTabWeekly: "📅 Mingguan",
+    missionTabRamadhan: "🌙 Ramadhan",
+    missionTabSyaban: "🌙 Sya'ban",
+    missionTabSeasonal: "🌙 Musiman",
+    missionEmptyCategory: "Belum ada misi di kategori ini.",
+    missionEmptySeasonalTitle: "Belum ada misi khusus bulan ini.",
+    missionEmptySeasonalDesc: "Nantikan misi spesial Ramadhan segera!",
+
+    // Mission Completion Options
+    missionCompletionAlone: "Sholat Sendiri",
+    missionCompletionCongregation: "Berjamaah di Masjid",
+
+    // Mission Hukum Labels
+    hukumWajib: "Wajib",
+    hukumSunnah: "Sunnah",
+    hukumMubah: "Mubah",
+    hukumMakruh: "Makruh",
+    hukumHaram: "Haram",
+
+    // Quran Browser
+    quranLastRead: "Terakhir Dibaca",
+    quranVerse: "Ayat",
+    quranNoHistory: "Belum ada riwayat",
+    quranStartReading: "Mulai membaca sekarang.",
+    quranBookmarks: "Tanda Baca",
+    quranVersesSaved: "ayat tersimpan",
+
+    // Bookmarks Page
+    bookmarksTitle: "Koleksi Tanda Baca",
+    bookmarksSubtitle: "Catatan dan ayat pilihan Anda",
+    bookmarksEmptyTitle: "Belum ada koleksi",
+    bookmarksEmptyDesc: "Jelajahi Al-Quran dan simpan ayat yang menyentuh hati Anda.",
+    bookmarksStartReading: "Mulai Membaca",
+    bookmarksLastReadLabel: "Terakhir Dibaca",
+    bookmarksSetLastRead: "Jadikan Terakhir Baca",
+    bookmarksCurrentlyReading: "Sedang Dibaca",
+    bookmarksDelete: "Hapus",
+    bookmarksDeleteConfirm: "Hapus tanda baca ini?",
+    bookmarksMarkedAsLastRead: "Ditandai sebagai Terakhir Baca 📖",
+    bookmarksEditTitle: "Edit Bookmark",
+    bookmarksNotePlaceholder: "Tulis catatan Anda di sini...",
+    bookmarksMarkAsLastRead: "Tandai Terakhir Dibaca",
+    bookmarksUpdateProgress: "Update progress membaca di halaman beranda",
+    bookmarksSave: "Simpan Perubahan",
+    quranSearchPlaceholder: "Cari Surat (e.g. Al-Kahf) atau Arti...",
+    quranSurahList: "Surat-Surat",
+    quranSurahCount: "SURAT",
+    quranVerseCount: "Ayat",
+    quranMakkah: "Mekah",
+    quranMadinah: "Madinah",
+    quranLoadError: "Gagal memuat data Al-Quran",
+    quranLoadErrorDesc: "Mohon periksa koneksi internet Anda dan coba lagi.",
+
+    // Quran Reader Settings
+    quranPlaySurah: "Putar Surat",
+    quranJumpToVerse: "Loncat ke Ayat",
+    quranJumpToVerseTitle: "Loncat ke Ayat",
+    quranGoToVerse: "Pergi ke Ayat",
+    quranSettingsTitle: "Pengaturan Tampilan",
+    quranModeRead: "Mode Baca",
+    quranModeList: "List",
+    quranModeMushaf: "Mushaf",
+    quranScriptType: "Jenis Teks Arab",
+    quranScriptStandard: "Standar Indonesia",
+    quranScriptTajweed: "Tajweed Berwarna",
+    quranOtherDisplay: "Tampilan Lainnya",
+    quranTransliteration: "Latin / Transliterasi",
+    quranFontSize: "Ukuran Huruf Arab",
+    quranVersesPerPage: "Ayat per Halaman",
+    quranSelectQari: "Pilih Qari",
+    quranPrevious: "Sebelumnya",
+    quranNext: "Selanjutnya",
+    quranPage: "Halaman",
+    quranOf: "/",
+
+    // Prayer Names
+    prayerImsak: "Imsak",
+    prayerFajr: "Subuh",
+    prayerSunrise: "Terbit",
+    prayerDhuhr: "Dzuhur",
+    prayerAsr: "Ashar",
+    prayerMaghrib: "Maghrib",
+    prayerIsha: "Isya",
+
+    // Quote of the Day
+    quoteText1: "Maka sesungguhnya bersama kesulitan ada kemudahan.",
+    quoteSource1: "QS. Al-Insyirah: 5",
+    quoteText2: "Dan berikanlah berita gembira kepada orang-orang yang sabar.",
+    quoteSource2: "QS. Al-Baqarah: 155",
+    quoteText3: "Allah tidak membebani seseorang melainkan sesuai dengan kesanggupannya.",
+    quoteSource3: "QS. Al-Baqarah: 286",
+    quoteText4: "Barangsiapa bertakwa kepada Allah niscaya Dia akan mengadakan baginya jalan keluar.",
+    quoteSource4: "QS. At-Talaq: 2",
+    quoteText5: "Dan Dia memberinya rezeki dari arah yang tiada disangka-sangkanya.",
+    quoteSource5: "QS. At-Talaq: 3",
+    quoteText6: "Ingatlah, hanya dengan mengingati Allah-lah hati menjadi tenteram.",
+    quoteSource6: "QS. Ar-Ra'd: 28",
+    quoteText7: "Sebaik-baik manusia adalah yang paling bermanfaat bagi manusia.",
+    quoteSource7: "HR. Ahmad",
+    quoteText8: "Senyummu di hadapan saudaramu adalah sedekah.",
+    quoteSource8: "HR. Tirmidzi",
+    quoteText9: "Janganlah kamu marah, maka bagimu surga.",
+    quoteSource9: "HR. Thabrani",
+    quoteText10: "Sholat itu adalah tiang agama.",
+    quoteSource10: "HR. Tirmidzi",
+
+    // Stats Page
+    statsPageTitle: "Statistik Ibadah",
+    statsPageSubtitle: "Pantau perkembangan spiritualmu",
+    statsWeekly: "Mingguan",
+    statsMonthly: "Bulanan",
+    statsTotalXP: "Total XP",
+    statsComingSoonFeature: "Fitur Sedang Dimatangkan ✨",
+    statsComingSoonMessage: "Kami sedang menyempurnakan algoritma statistik agar jurnal ibadah Anda lebih akurat dan bermakna. Sabar ya, Kak!",
+    statsBackHome: "Kembali ke Beranda",
+
+    // Nawaetu AI Page
+    tanyaTitle: "Nawaetu AI",
+    tanyaSubtitle: "Asisten Muslim Digital",
+    tanyaGreetingMorning: "Assalamu'alaikum kak! 🙏 Semoga hari ini penuh berkah. Ada yang ingin ditanyakan?",
+    tanyaGreetingDay: "Assalamu'alaikum kak! ☀️ Selamat pagi! Aku siap menemani diskusi spiritualmu.",
+    tanyaGreetingNoon: "Assalamu'alaikum kak! 🌤️ Siang ini ada yang mau diceritakan?",
+    tanyaGreetingAfternoon: "Assalamu'alaikum kak! 🌅 Sore yang indah. Yuk ngobrol seputar ibadah.",
+    tanyaGreetingEvening: "Assalamu'alaikum kak! 🌙 Selamat malam. Mau curhat atau tanya hukum islam?",
+    tanyaPlaceholder: "Tanya sesuatu...",
+    tanyaSend: "Kirim",
+    tanyaDailyLimit: "Batas Pertanyaan",
+    tanyaLimitReached: "Kamu sudah mencapai batas 5 pertanyaan hari ini.",
+    tanyaLimitReset: "Limit akan reset besok jam 00:00",
+    tanyaUnlockPremium: "Unlimited untuk Muhsinin",
+    tanyaQuickPrompts: "Prompt Cepat",
+    tanyaPrompt1: "Saya merasa malas sholat...",
+    tanyaPrompt2: "Sedang banyak pikiran",
+    tanyaPrompt3: "Tips khusyuk?",
+    tanyaPrompt4: "Apa keutamaan subuh?",
+
+    // Toast Messages
+    toastMissionComplete: "Misi Selesai!",
+    toastMissionMsg1: "Masya Allah! Istiqamah terus ya ✨",
+    toastMissionMsg2: "Alhamdulillah, satu kebaikan tercatat 🤲",
+    toastMissionMsg3: "Barakallah! Semangat ibadahnya 🔥",
+    toastMissionMsg4: "Keren! Niat lurus, amal mulus 💎",
+
+    // Onboarding
+    onboardingSkip: "Lewati Intro",
+    onboardingBack: "Kembali",
+    onboardingNext: "Lanjut",
+    onboardingFinish: "Selesai & Mulai 🚀",
+    onboardingNameTitle: "Siapa namamu?",
+    onboardingNameDesc: "Agar Nawaetu AI bisa menyapamu dengan akrab.",
+    onboardingNamePlaceholder: "Nama Panggilan",
+    onboardingDefaultName: "Sobat Nawaetu",
+    onboardingGenderTitle: "Apa Jenis Kelaminmu?",
+    onboardingGenderDesc: "Untuk menyesuaikan fiqih ibadah & tema.",
+    onboardingMaleLabel: "Laki-laki",
+    onboardingMaleSub: "Tema Biru, Fiqih Pria",
+    onboardingFemaleLabel: "Perempuan",
+    onboardingFemaleSub: "Tema Pink, Fiqih Wanita",
+    onboardingArchetypeTitle: "Tipe Pejuang Ibadah?",
+    onboardingArchetypeDesc: "Menentukan target misi harianmu.",
+    onboardingArchetypeBeginnerLabel: "Beginner",
+    onboardingArchetypeBeginnerDesc: "Fokus Ibadah Wajib",
+    onboardingArchetypeStriverLabel: "Striver",
+    onboardingArchetypeStriverDesc: "Wajib + Sunnah Rutin",
+    onboardingArchetypeDedicatedLabel: "Dedicated",
+    onboardingArchetypeDedicatedDesc: "Target Ibadah Tinggi",
+
+    // Onboarding Slides
+    onboardingSlide1Title: "Tanpa Install, Hemat Memori",
+    onboardingSlide1Desc: "Tidak perlu download dari PlayStore/AppStore. Cukup akses web ini, ringan dan cepat.",
+    onboardingSlide1Hint: "Hint: Tambahkan ke Home Screen untuk akses instan.",
+    onboardingSlide2Title: "Al-Qur'an & Mentor AI Terpercaya",
+    onboardingSlide2Desc: "Baca Al-Qur'an, tafsir, dan tanya jawab agama. Semua konten AI dipastikan bersumber dari Al-Qur'an & Hadits Shahih.",
+    onboardingSlide2Hint: "Tenang, sumbernya valid & dapat dipertanggungjawabkan.",
+    onboardingSlide3Title: "Gamifikasi Personal",
+    onboardingSlide3Desc: "Misi harian & target XP akan disesuaikan dengan profil Anda (Gender & Tipe Pejuang) agar lebih relevan.",
+    onboardingSlide3Hint: "Ibadah jadi lebih semangat dengan target yang pas.",
+    onboardingSlide4Title: "Tanpa Iklan & Privasi Terjaga",
+    onboardingSlide4Desc: "Fokus ibadah tanpa gangguan iklan. Data tersimpan lokal di HP Anda, tidak kami jual.",
+    onboardingSlide4Hint: "PENTING: Jangan 'Clear Cache' agar data tidak hilang.",
+};
